Split schema compare internals by concern
