+++ migration/schemadiff/internal/compare/compare.go
@@ -1,3362 +1,2 @@
+// Package compare contains concern-level schema diff comparators.
 package compare
-
-import (
-	"fmt"
-	"regexp"
-	"slices"
-	"sort"
-	"strings"
-	"unicode"
-
-	"github.com/stokaro/ptah/config"
-	"github.com/stokaro/ptah/core/convert/fromschema"
-	"github.com/stokaro/ptah/core/goschema"
-	"github.com/stokaro/ptah/dbschema/types"
-	"github.com/stokaro/ptah/migration/schemadiff/internal/normalize"
-	difftypes "github.com/stokaro/ptah/migration/schemadiff/types"
-	"github.com/stokaro/ptah/migration/typechange"
-)
-
-// Regular expressions for constraint-based index detection
-var (
-	// PostgreSQL constraint-based unique index pattern: tablename_columnname_key
-	postgresConstraintPattern = regexp.MustCompile(`^[a-zA-Z_][a-zA-Z0-9_]*_[a-zA-Z_][a-zA-Z0-9_]*_key$`)
-
-	// MySQL/MariaDB constraint-based unique index patterns
-	mysqlUKPattern           = regexp.MustCompile(`^uk_[a-zA-Z_][a-zA-Z0-9_]*`)
-	mysqlTableColumnsPattern = regexp.MustCompile(`^[a-zA-Z_][a-zA-Z0-9_]*_[a-zA-Z_][a-zA-Z0-9_]*$`)
-
-	// Custom index patterns (these should NOT be considered constraint-based)
-	// Match indexes that start with "idx_" or "index_", or end with "_idx" or "_index"
-	customIndexPattern = regexp.MustCompile(`(?i)(^(idx|index)_|_(idx|index)$)`)
-
-	defaultAggregateAliasPattern = regexp.MustCompile(`\b(count|sum|avg|min|max)\(([^)]*)\)\s+as\s+([a-z_][a-z0-9_]*)\b`)
-	schemaQualifierPattern       = regexp.MustCompile(`\b[a-z_][a-z0-9_]*\.`)
-)
-
-// TablesAndColumns performs comprehensive table and column comparison between generated and database schemas.
-//
-// This function is the core table comparison engine that identifies structural differences
-// between the target schema (from Go struct annotations) and the current database schema.
-// It handles table additions, removals, and detailed column-level modifications.
-//
-// # Comparison Process
-//
-// The function performs comparison in three phases:
-//  1. **Table Discovery**: Creates lookup maps for efficient table comparison
-//  2. **Table Diff Analysis**: Identifies added and removed tables
-//  3. **Column Comparison**: For existing tables, performs detailed column analysis
-//
-// # Algorithm Complexity
-//
-// - Time Complexity: O(n + m + k) where n=generated tables, m=database tables, k=total columns
-// - Space Complexity: O(n + m) for lookup maps
-// - Optimized for large schemas with efficient map-based lookups
-//
-// # Embedded Field Handling
-//
-// The function properly handles embedded fields by delegating to TableColumns(),
-// which processes embedded fields through the transform package to ensure generated
-// fields are correctly compared against database columns.
-//
-// # Example Scenarios
-//
-// **New table detection**:
-//   - Generated schema has "users" table
-//   - Database schema doesn't have "users" table
-//   - Result: "users" added to diff.TablesAdded
-//
-// **Removed table detection**:
-//   - Database has "legacy_data" table
-//   - Generated schema doesn't define "legacy_data"
-//   - Result: "legacy_data" added to diff.TablesRemoved
-//
-// **Modified table detection**:
-//   - Both schemas have "products" table
-//   - Column structures differ (new columns, type changes, etc.)
-//   - Result: TableDiff added to diff.TablesModified
-//
-// # Parameters
-//
-//   - generated: Target schema parsed from Go struct annotations
-//   - database: Current database schema from executor introspection
-//   - diff: SchemaDiff structure to populate with discovered differences
-//
-// # Side Effects
-//
-// Modifies the provided diff parameter by populating:
-//   - diff.TablesAdded: Tables that need to be created
-//   - diff.TablesRemoved: Tables that exist in database but not in target schema
-//   - diff.TablesModified: Tables with structural differences
-//
-// # Output Consistency
-//
-// Results are sorted alphabetically for consistent output across multiple runs,
-// ensuring deterministic migration generation and reliable testing.
-func TablesAndColumns(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff) {
-	// Create maps for quick lookup
-	genTables := make(map[string]goschema.Table)
-	for _, table := range generated.Tables {
-		genTables[table.QualifiedName()] = table
-	}
-
-	dbTables := make(map[string]types.DBTable)
-	for _, table := range database.Tables {
-		dbTables[table.QualifiedName()] = table
-	}
-
-	// Find added and removed tables
-	for tableName := range genTables {
-		if _, exists := dbTables[tableName]; !exists {
-			diff.TablesAdded = append(diff.TablesAdded, tableName)
-		}
-	}
-
-	for tableName := range dbTables {
-		if _, exists := genTables[tableName]; !exists {
-			diff.TablesRemoved = append(diff.TablesRemoved, tableName)
-		}
-	}
-
-	// Find modified tables (compare columns)
-	for tableName, genTable := range genTables {
-		if dbTable, exists := dbTables[tableName]; exists {
-			tableDiff := TableColumns(genTable, dbTable, generated)
-			if len(tableDiff.ColumnsAdded) > 0 || len(tableDiff.ColumnsRemoved) > 0 || len(tableDiff.ColumnsModified) > 0 {
-				diff.TablesModified = append(diff.TablesModified, tableDiff)
-			}
-		}
-	}
-
-	// Sort for consistent output
-	sort.Strings(diff.TablesAdded)
-	sort.Strings(diff.TablesRemoved)
-	sort.Slice(diff.TablesModified, func(i, j int) bool {
-		return diff.TablesModified[i].TableName < diff.TablesModified[j].TableName
-	})
-}
-
-// TableColumns performs detailed column-level comparison within a specific table.
-//
-// This function is responsible for the complex task of comparing column structures
-// between a generated table definition and an existing database table. It handles
-// embedded field processing, column mapping, and detailed property comparison.
-//
-// # Embedded Field Processing
-//
-// The function's most complex aspect is handling embedded fields:
-//  1. **Field Expansion**: Uses transform.ProcessEmbeddedFields() to expand embedded structs
-//  2. **Field Combination**: Merges original fields with embedded-generated fields
-//  3. **Struct Filtering**: Only processes fields belonging to the target struct
-//
-// This ensures that embedded fields (like timestamps, audit info) are properly
-// compared against their corresponding database columns.
-//
-// # Comparison Algorithm
-//
-// The function performs comparison in three phases:
-//  1. **Column Discovery**: Creates lookup maps for efficient column comparison
-//  2. **Addition/Removal Detection**: Identifies new and removed columns
-//  3. **Modification Analysis**: Compares properties of existing columns
-//
-// # Example Scenarios
-//
-// **Embedded field handling**:
-//
-//	```go
-//	type User struct {
-//	    ID   int    `db:"id"`
-//	    Name string `db:"name"`
-//	    Timestamps // Embedded struct with CreatedAt, UpdatedAt
-//	}
-//	```
-//	The function expands Timestamps fields and compares them against database columns.
-//
-// **Column addition detection**:
-//   - Generated schema has "email" column
-//   - Database table doesn't have "email" column
-//   - Result: "email" added to TableDiff.ColumnsAdded
-//
-// **Column modification detection**:
-//   - Both have "name" column
-//   - Generated: VARCHAR(255), Database: VARCHAR(100)
-//   - Result: ColumnDiff added to TableDiff.ColumnsModified
-//
-// # Parameters
-//
-//   - genTable: Generated table definition from Go struct annotations
-//   - dbTable: Current database table structure from introspection
-//   - generated: Complete parse result containing all fields and embedded field definitions
-//
-// # Return Value
-//
-// Returns a TableDiff containing:
-//   - ColumnsAdded: New columns that need to be added
-//   - ColumnsRemoved: Existing columns that should be removed
-//   - ColumnsModified: Columns with property differences
-//
-// # Performance Considerations
-//
-// - Time Complexity: O(n + m + k) where n=generated columns, m=database columns, k=embedded fields
-// - Space Complexity: O(n + m) for lookup maps
-// - Embedded field processing adds overhead but is necessary for accurate comparison
-//
-// # Output Consistency
-//
-// Column lists are sorted alphabetically for deterministic output and reliable testing.
-func TableColumns(genTable goschema.Table, dbTable types.DBTable, generated *goschema.Database) difftypes.TableDiff {
-	tableDiff := difftypes.TableDiff{TableName: genTable.QualifiedName()}
-
-	// Process embedded fields to get the complete field list (same as generators do)
-	embeddedGeneratedFields := processEmbeddedFieldsForStruct(generated.EmbeddedFields, generated.Fields, genTable.StructName)
-
-	// Combine original fields with embedded-generated fields
-	allFields := append(([]goschema.Field)(nil), generated.Fields...)
-	allFields = append(allFields, embeddedGeneratedFields...)
-
-	// Create maps for quick lookup
-	genColumns := make(map[string]goschema.Field)
-	for _, field := range allFields {
-		if field.StructName == genTable.StructName {
-			genColumns[field.Name] = field
-		}
-	}
-
-	dbColumns := make(map[string]types.DBColumn)
-	for _, col := range dbTable.Columns {
-		dbColumns[col.Name] = col
-	}
-
-	// Find added and removed columns
-	for colName := range genColumns {
-		if _, exists := dbColumns[colName]; !exists {
-			tableDiff.ColumnsAdded = append(tableDiff.ColumnsAdded, colName)
-		}
-	}
-
-	for colName := range dbColumns {
-		if _, exists := genColumns[colName]; !exists {
-			tableDiff.ColumnsRemoved = append(tableDiff.ColumnsRemoved, colName)
-		}
-	}
-
-	// Find modified columns
-	for colName, genCol := range genColumns {
-		if dbCol, exists := dbColumns[colName]; exists {
-			if columnInTablePrimaryKey(genTable, genCol.Name) {
-				genCol = normalizeTablePrimaryKeyColumn(genCol, dbCol)
-			}
-			colDiff := Columns(genCol, dbCol)
-			if len(colDiff.Changes) > 0 {
-				tableDiff.ColumnsModified = append(tableDiff.ColumnsModified, colDiff)
-			}
-		}
-	}
-
-	// Sort for consistent output
-	sort.Strings(tableDiff.ColumnsAdded)
-	sort.Strings(tableDiff.ColumnsRemoved)
-	sort.Slice(tableDiff.ColumnsModified, func(i, j int) bool {
-		return tableDiff.ColumnsModified[i].ColumnName < tableDiff.ColumnsModified[j].ColumnName
-	})
-
-	return tableDiff
-}
-
-// Columns performs detailed property-level comparison between a generated column and database column.
-//
-// This function is the most granular level of schema comparison, analyzing individual
-// column properties to detect differences that require migration. It handles complex
-// cross-database type normalization and property comparison logic.
-//
-// # Property Comparison Categories
-//
-// The function compares five main categories of column properties:
-//  1. **Data Types**: Handles cross-database type normalization and comparison
-//  2. **Nullability**: Considers primary key implications and explicit nullable settings
-//  3. **Primary Key**: Compares primary key constraint status
-//  4. **Uniqueness**: Compares unique constraint status
-//  5. **Default Values**: Handles auto-increment special cases and type-specific normalization
-//
-// # Complex Logic Areas
-//
-// **Type Normalization**:
-//   - Uses Type() to handle cross-database type variations
-//   - Considers both DataType and UDTName from database introspection
-//   - Handles PostgreSQL user-defined types vs standard types
-//
-// **Nullability Logic**:
-//   - Primary key columns are always NOT NULL regardless of field definition
-//   - Explicit nullable settings override default behavior
-//   - Database "YES"/"NO" strings converted to boolean for comparison
-//
-// **Auto-increment Handling**:
-//   - SERIAL columns have special default value handling
-//   - Database shows sequence defaults, but entities expect empty defaults
-//   - Prevents false positives for auto-increment columns
-//
-// # Example Comparisons
-//
-// **Type difference detection**:
-//
-//	```
-//	Generated: VARCHAR(255)
-//	Database:  VARCHAR(100)
-//	Result:    Changes["type"] = "varchar -> varchar" (normalized)
-//	```
-//
-// **Nullability change**:
-//
-//	```
-//	Generated: nullable=false
-//	Database:  nullable=true
-//	Result:    Changes["nullable"] = "true -> false"
-//	```
-//
-// **Primary key promotion**:
-//
-//	```
-//	Generated: primary=true
-//	Database:  primary=false
-//	Result:    Changes["primary_key"] = "false -> true"
-//	```
-//
-// **Default value normalization**:
-//
-//	```
-//	Generated: default=""
-//	Database:  default_expr="NULL"
-//	Result:    No change (both normalize to empty string)
-//	```
-//
-// # Parameters
-//
-//   - genCol: Generated column definition from Go struct field
-//   - dbCol: Current database column from introspection
-//
-// # Return Value
-//
-// Returns a ColumnDiff with:
-//   - ColumnName: Name of the column being compared
-//   - Changes: Map of property changes in "old -> new" format
-//
-// # Cross-Database Considerations
-//
-// The function handles database-specific variations:
-//   - **PostgreSQL**: UDT names, SERIAL types, native boolean types
-//   - **MySQL/MariaDB**: TINYINT boolean representation, AUTO_INCREMENT
-//   - **Type mapping**: Intelligent normalization for accurate comparison
-func Columns(genCol goschema.Field, dbCol types.DBColumn) difftypes.ColumnDiff {
-	colDiff := difftypes.ColumnDiff{
-		ColumnName: genCol.Name,
-		Changes:    make(map[string]string),
-	}
-
-	// Legacy ClickHouse-only guard: old goschema models cannot express
-	// MATERIALIZED / ALIAS / EPHEMERAL columns. Once the schema side carries a
-	// generated expression, compare it normally below.
-	if dbCol.GeneratedKind != "" && genCol.GeneratedExpression == "" {
-		return colDiff
-	}
-
-	// Compare data types (simplified)
-	genType := normalize.Type(genCol.Type)
-	dbRawType := rawDBColumnType(dbCol)
-	dbType := normalize.Type(dbRawType)
-
-	if genType != dbType {
-		colDiff.Changes["type"] = fmt.Sprintf("%s -> %s", dbType, genType)
-	} else if typechange.IsNarrowing(dbRawType, genCol.Type) {
-		colDiff.Changes["type"] = fmt.Sprintf("%s -> %s", dbRawType, genCol.Type)
-	}
-
-	// Compare nullable (primary keys are always NOT NULL regardless of the field definition)
-	genNullable := genCol.Nullable
-	if genCol.Primary {
-		genNullable = false // Primary keys are always NOT NULL
-	}
-	dbNullable := dbCol.IsNullable == "YES"
-	if genNullable != dbNullable {
-		colDiff.Changes["nullable"] = fmt.Sprintf("%t -> %t", dbNullable, genNullable)
-	}
-
-	// Compare primary key
-	genPrimary := genCol.Primary
-	dbPrimary := dbCol.IsPrimaryKey
-	if genPrimary != dbPrimary {
-		colDiff.Changes["primary_key"] = fmt.Sprintf("%t -> %t", dbPrimary, genPrimary)
-	}
-
-	// Compare unique
-	genUnique := genCol.Unique
-	dbUnique := dbCol.IsUnique
-	if genUnique != dbUnique {
-		colDiff.Changes["unique"] = fmt.Sprintf("%t -> %t", dbUnique, genUnique)
-	}
-	if diff := generatedColumnDiff(genCol, dbCol); diff != "" {
-		colDiff.Changes["generated"] = diff
-	}
-
-	// Compare default values (simplified)
-	genDefault := genCol.Default
-	if genDefault == "" {
-		genDefault = genCol.DefaultExpr
-	}
-	dbDefault := ""
-	if dbCol.ColumnDefault != nil {
-		dbDefault = *dbCol.ColumnDefault
-	}
-
-	// For auto-increment/SERIAL columns, ignore default value differences
-	// because the database will show the sequence default but the entity expects empty
-	isAutoIncrement := dbCol.IsAutoIncrement || strings.Contains(strings.ToUpper(genCol.Type), "SERIAL")
-	if !isAutoIncrement {
-		normalizedDbDefault := normalize.DefaultValue(dbDefault, dbType)
-
-		idxName := "default"
-		if normalize.IsDefaultExpr(dbDefault) {
-			idxName = "default_expr"
-		}
-
-		normalizeGenDefaultFn := normalize.DefaultValue(genDefault, "")
-
-		if normalizeGenDefaultFn != normalizedDbDefault {
-			colDiff.Changes[idxName] = fmt.Sprintf("%s -> %s", dbDefault, genDefault)
-		}
-	}
-
-	return colDiff
-}
-
-func normalizeTablePrimaryKeyColumn(genCol goschema.Field, dbCol types.DBColumn) goschema.Field {
-	genCol.Nullable = false
-	genCol.Primary = dbCol.IsPrimaryKey
-	return genCol
-}
-
-func columnInTablePrimaryKey(table goschema.Table, column string) bool {
-	return slices.Contains(tablePrimaryKeyColumns(table), column)
-}
-
-func tablePrimaryKeyColumns(table goschema.Table) []string {
-	if len(table.PrimaryKeyParts) == 0 {
-		return nonEmptyNames(table.PrimaryKey)
-	}
-	columns := make([]string, 0, len(table.PrimaryKeyParts))
-	for _, part := range table.PrimaryKeyParts {
-		if name := strings.TrimSpace(part.Name); name != "" {
-			columns = append(columns, name)
-		}
-	}
-	return columns
-}
-
-func nonEmptyNames(names []string) []string {
-	filtered := make([]string, 0, len(names))
-	for _, name := range names {
-		if name := strings.TrimSpace(name); name != "" {
-			filtered = append(filtered, name)
-		}
-	}
-	return filtered
-}
-
-func generatedColumnDiff(genCol goschema.Field, dbCol types.DBColumn) string {
-	genExpr := strings.TrimSpace(genCol.GeneratedExpression)
-	dbExpr := ""
-	if dbCol.GeneratedExpression != nil {
-		dbExpr = strings.TrimSpace(*dbCol.GeneratedExpression)
-	}
-	genKind := strings.ToUpper(strings.TrimSpace(genCol.GeneratedKind))
-	dbKind := strings.ToUpper(strings.TrimSpace(dbCol.GeneratedKind))
-	if genExpr == dbExpr && genKind == dbKind {
-		return ""
-	}
-	return fmt.Sprintf("%s %s -> %s %s", dbKind, dbExpr, genKind, genExpr)
-}
-
-func rawDBColumnType(dbCol types.DBColumn) string {
-	rawType := strings.TrimSpace(dbCol.ColumnType)
-	if rawType == "" && dbCol.UDTName != "" {
-		rawType = strings.TrimSpace(dbCol.UDTName)
-	}
-	if rawType == "" {
-		rawType = strings.TrimSpace(dbCol.DataType)
-	}
-
-	if strings.Contains(rawType, "(") {
-		return rawType
-	}
-	switch normalize.Type(rawType) {
-	case "varchar":
-		if dbCol.CharacterMaxLength != nil {
-			return fmt.Sprintf("%s(%d)", rawType, *dbCol.CharacterMaxLength)
-		}
-	case "decimal":
-		if dbCol.NumericPrecision == nil {
-			return rawType
-		}
-		if dbCol.NumericScale != nil {
-			return fmt.Sprintf("%s(%d,%d)", rawType, *dbCol.NumericPrecision, *dbCol.NumericScale)
-		}
-		return fmt.Sprintf("%s(%d)", rawType, *dbCol.NumericPrecision)
-	}
-	return rawType
-}
-
-// SearchColumnByName searches for a specific column difference by name within a slice of column diffs.
-//
-// This utility function provides efficient lookup of column differences by name, which is
-// commonly needed when processing migration results or analyzing specific column changes.
-// It performs a linear search through the provided slice and returns a pointer to the
-// first matching ColumnDiff.
-//
-// # Search Algorithm
-//
-// The function uses a simple linear search with O(n) time complexity:
-//  1. **Iteration**: Loops through each ColumnDiff in the provided slice
-//  2. **Name Matching**: Compares ColumnName field with the target column name
-//  3. **Early Return**: Returns immediately upon finding the first match
-//  4. **Not Found**: Returns nil if no matching column is found
-//
-// # Use Cases
-//
-// **Migration Analysis**:
-//   - Check if a specific column has changes before generating migration SQL
-//   - Retrieve detailed change information for a particular column
-//   - Validate expected changes in automated tests
-//
-// **Conditional Processing**:
-//   - Apply different migration strategies based on specific column changes
-//   - Skip certain operations if particular columns are not modified
-//   - Generate warnings for potentially dangerous column modifications
-//
-// # Example Usage
-//
-// **Finding a specific column change**:
-//
-//	```go
-//	tableDiff := compare.TableColumns(genTable, dbTable, generated)
-//	emailDiff := compare.ColumnByName(tableDiff.ColumnsModified, "email")
-//	if emailDiff != nil {
-//	    if _, hasTypeChange := emailDiff.Changes["type"]; hasTypeChange {
-//	        log.Println("Email column type is changing")
-//	    }
-//	}
-//	```
-//
-// **Validation in tests**:
-//
-//	```go
-//	result := compare.TableColumns(genTable, dbTable, generated)
-//	nameDiff := compare.ColumnByName(result.ColumnsModified, "name")
-//	assert.NotNil(t, nameDiff)
-//	assert.Equal(t, "varchar -> text", nameDiff.Changes["type"])
-//	```
-//
-// **Conditional migration logic**:
-//
-//	```go
-//	for _, tableDiff := range schemaDiff.TablesModified {
-//	    if pkDiff := compare.ColumnByName(tableDiff.ColumnsModified, "id"); pkDiff != nil {
-//	        if _, hasPKChange := pkDiff.Changes["primary_key"]; hasPKChange {
-//	            // Handle primary key changes with special care
-//	            generatePrimaryKeyMigration(tableDiff.TableName, pkDiff)
-//	        }
-//	    }
-//	}
-//	```
-//
-// # Parameters
-//
-//   - diffs: Slice of ColumnDiff structures to search through
-//   - columnName: Name of the column to find (case-sensitive exact match)
-//
-// # Return Value
-//
-// Returns a pointer to the first ColumnDiff with matching ColumnName, or nil if not found.
-// The returned pointer references the original ColumnDiff in the slice, so modifications
-// will affect the original data structure.
-//
-// # Performance Considerations
-//
-// - Time Complexity: O(n) where n is the number of column diffs
-// - Space Complexity: O(1) - no additional memory allocation
-// - For large numbers of columns, consider using a map-based lookup if called frequently
-//
-// # Thread Safety
-//
-// This function is read-only and thread-safe when used concurrently on the same data.
-// However, if the underlying slice is being modified concurrently, appropriate
-// synchronization is required.
-//
-// # Edge Cases
-//
-// - Empty slice: Returns nil immediately
-// - Nil slice: Returns nil immediately (no panic)
-// - Duplicate column names: Returns the first match encountered
-// - Case sensitivity: Performs exact string matching (case-sensitive)
-func SearchColumnByName(diffs []difftypes.ColumnDiff, columnName string) *difftypes.ColumnDiff {
-	for _, diff := range diffs {
-		if diff.ColumnName == columnName {
-			return &diff
-		}
-	}
-	return nil
-}
-
-// Enums performs comprehensive enum type comparison between generated and database schemas.
-//
-// This function handles the comparison of enum type definitions, which is particularly
-// complex due to database-specific enum implementations and the challenges of enum
-// value modification across different database systems.
-//
-// # Database-Specific Enum Handling
-//
-// **PostgreSQL**:
-//   - Native ENUM types with CREATE TYPE statements
-//   - Supports adding enum values but not removing them easily
-//   - Enum values are stored in system catalogs
-//
-// **MySQL/MariaDB**:
-//   - Inline ENUM syntax in column definitions
-//   - Supports both adding and removing enum values
-//   - Enum values are part of column type definition
-//
-// **SQLite**:
-//   - No native enum support
-//   - Uses CHECK constraints for enum-like behavior
-//
-// # Comparison Algorithm
-//
-// The function performs comparison in three phases:
-//  1. **Enum Discovery**: Creates lookup maps for efficient enum comparison
-//  2. **Addition/Removal Detection**: Identifies new and removed enum types
-//  3. **Value Modification Analysis**: Compares enum values for existing types
-//
-// # Example Scenarios
-//
-// **New enum detection**:
-//   - Generated schema defines "status_type" enum
-//   - Database doesn't have "status_type" enum
-//   - Result: "status_type" added to diff.EnumsAdded
-//
-// **Enum value addition**:
-//   - Both have "priority_level" enum
-//   - Generated: ["low", "medium", "high", "critical"]
-//   - Database: ["low", "medium", "high"]
-//   - Result: EnumDiff with ValuesAdded=["critical"]
-//
-// **Enum value removal** (problematic):
-//   - Generated: ["active", "inactive"]
-//   - Database: ["active", "inactive", "deprecated"]
-//   - Result: EnumDiff with ValuesRemoved=["deprecated"]
-//   - Note: May require manual intervention in PostgreSQL
-//
-// # Parameters
-//
-//   - generated: Target schema parsed from Go struct annotations
-//   - database: Current database schema from executor introspection
-//   - diff: SchemaDiff structure to populate with discovered differences
-//
-// # Side Effects
-//
-// Modifies the provided diff parameter by populating:
-//   - diff.EnumsAdded: Enum types that need to be created
-//   - diff.EnumsRemoved: Enum types that exist in database but not in target schema
-//   - diff.EnumsModified: Enum types with value differences
-//
-// # Migration Considerations
-//
-// Enum modifications can be complex:
-//   - Adding values is generally safe
-//   - Removing values may require data migration
-//   - PostgreSQL enum removal requires recreating the enum type
-//   - MySQL enum changes require ALTER TABLE statements
-//
-// # Output Consistency
-//
-// Results are sorted alphabetically for consistent output across multiple runs.
-func Enums(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff) {
-	// Create maps for quick lookup
-	genEnums := make(map[string]goschema.Enum)
-	for _, enum := range generated.Enums {
-		genEnums[enum.Name] = enum
-	}
-
-	dbEnums := make(map[string]types.DBEnum)
-	for _, enum := range database.Enums {
-		dbEnums[enum.Name] = enum
-	}
-
-	// Find added and removed enums
-	for enumName := range genEnums {
-		if _, exists := dbEnums[enumName]; !exists {
-			diff.EnumsAdded = append(diff.EnumsAdded, enumName)
-		}
-	}
-
-	for enumName := range dbEnums {
-		if _, exists := genEnums[enumName]; !exists {
-			diff.EnumsRemoved = append(diff.EnumsRemoved, enumName)
-		}
-	}
-
-	// Find modified enums
-	for enumName, genEnum := range genEnums {
-		if dbEnum, exists := dbEnums[enumName]; exists {
-			enumDiff := EnumValues(genEnum, dbEnum)
-			if len(enumDiff.ValuesAdded) > 0 || len(enumDiff.ValuesRemoved) > 0 {
-				diff.EnumsModified = append(diff.EnumsModified, enumDiff)
-			}
-		}
-	}
-
-	// Sort for consistent output
-	sort.Strings(diff.EnumsAdded)
-	sort.Strings(diff.EnumsRemoved)
-	sort.Slice(diff.EnumsModified, func(i, j int) bool {
-		return diff.EnumsModified[i].EnumName < diff.EnumsModified[j].EnumName
-	})
-}
-
-// EnumValues performs detailed value-level comparison between generated and database enum types.
-//
-// This function analyzes the specific values within an enum type to determine what
-// changes are needed to bring the database enum in line with the generated enum
-// definition. It uses set-based comparison for efficient value difference detection.
-//
-// # Algorithm Details
-//
-// The function uses a set-based approach for optimal performance:
-//  1. **Set Creation**: Converts value slices to boolean maps for O(1) lookup
-//  2. **Addition Detection**: Finds values in generated enum but not in database
-//  3. **Removal Detection**: Finds values in database enum but not in generated
-//  4. **Result Sorting**: Ensures deterministic output for consistent migrations
-//
-// # Example Scenarios
-//
-// **Value addition**:
-//
-//	```
-//	Generated: ["draft", "published", "archived", "deleted"]
-//	Database:  ["draft", "published", "archived"]
-//	Result:    ValuesAdded=["deleted"], ValuesRemoved=[]
-//	```
-//
-// **Value removal**:
-//
-//	```
-//	Generated: ["active", "inactive"]
-//	Database:  ["active", "inactive", "deprecated", "legacy"]
-//	Result:    ValuesAdded=[], ValuesRemoved=["deprecated", "legacy"]
-//	```
-//
-// **Mixed changes**:
-//
-//	```
-//	Generated: ["pending", "approved", "rejected", "canceled"]
-//	Database:  ["pending", "approved", "denied"]
-//	Result:    ValuesAdded=["rejected", "canceled"], ValuesRemoved=["denied"]
-//	```
-//
-// # Performance Characteristics
-//
-// - Time Complexity: O(n + m) where n=generated values, m=database values
-// - Space Complexity: O(n + m) for the boolean maps
-// - Optimized for large enum value sets with efficient set operations
-//
-// # Parameters
-//
-//   - genEnum: Generated enum definition from Go struct annotations
-//   - dbEnum: Current database enum from introspection
-//
-// # Return Value
-//
-// Returns an EnumDiff containing:
-//   - EnumName: Name of the enum being compared
-//   - ValuesAdded: Values that need to be added to the database enum
-//   - ValuesRemoved: Values that exist in database but not in generated enum
-//
-// # Migration Implications
-//
-// **Adding values**: Generally safe operation across all databases
-// **Removing values**: May require careful consideration:
-//   - Check if removed values are used in existing data
-//   - PostgreSQL requires enum recreation for value removal
-//   - MySQL allows value removal but may affect existing data
-//
-// # Output Consistency
-//
-// Value lists are sorted alphabetically to ensure deterministic migration
-// generation and reliable testing across multiple runs.
-func EnumValues(genEnum goschema.Enum, dbEnum types.DBEnum) difftypes.EnumDiff {
-	enumDiff := difftypes.EnumDiff{EnumName: genEnum.Name}
-
-	// Create sets for comparison
-	genValues := make(map[string]bool)
-	for _, value := range genEnum.Values {
-		genValues[value] = true
-	}
-
-	dbValues := make(map[string]bool)
-	for _, value := range dbEnum.Values {
-		dbValues[value] = true
-	}
-
-	// Find added and removed values
-	for value := range genValues {
-		if !dbValues[value] {
-			enumDiff.ValuesAdded = append(enumDiff.ValuesAdded, value)
-		}
-	}
-
-	for value := range dbValues {
-		if !genValues[value] {
-			enumDiff.ValuesRemoved = append(enumDiff.ValuesRemoved, value)
-		}
-	}
-
-	// Sort for consistent output
-	sort.Strings(enumDiff.ValuesAdded)
-	sort.Strings(enumDiff.ValuesRemoved)
-
-	return enumDiff
-}
-
-// isConstraintBasedUniqueIndex determines if a unique index was automatically created by a UNIQUE constraint.
-//
-// Different database systems create unique indexes with different naming patterns when UNIQUE
-// constraints are defined on columns:
-//
-// **PostgreSQL**:
-//   - tablename_columnname_key (single column)
-//   - tablename_columnname1_columnname2_key (multiple columns)
-//
-// **MySQL/MariaDB**:
-//   - Simple column names (e.g., "email", "username") for single-column constraints
-//   - Constraint names for multi-column constraints (e.g., "uk_users_email_name")
-//
-// This function identifies such constraint-based indexes to distinguish them from explicitly
-// defined unique indexes created via schema annotations.
-//
-// # Assumptions
-//
-// This function relies on standard naming conventions used by database systems for
-// constraint-based indexes. These patterns may vary with different database versions,
-// configurations, or custom naming schemes. The detection is based on common patterns
-// observed in PostgreSQL 12+, MySQL 8.0+, and MariaDB 10.5+.
-//
-// # Parameters
-//
-//   - indexName: The name of the index to check
-//   - tableName: The name of the table the index belongs to
-//   - columns: The columns that the index covers (used for MySQL/MariaDB detection)
-//
-// # Returns
-//
-// Returns true if the index appears to be constraint-based, false if it's explicitly defined.
-//
-// # Examples
-//
-//	// PostgreSQL
-//	isConstraintBasedUniqueIndex("users_email_key", "users", []string{"email"})     // true
-//	isConstraintBasedUniqueIndex("tenants_slug_idx", "tenants", []string{"slug"})   // false
-//
-//	// MySQL/MariaDB
-//	isConstraintBasedUniqueIndex("email", "users", []string{"email"})               // true
-//	isConstraintBasedUniqueIndex("idx_users_custom", "users", []string{"email"})    // false
-func isConstraintBasedUniqueIndex(indexName, tableName string, columns []string) bool {
-	// PostgreSQL pattern: tablename_columnname_key
-	if strings.HasSuffix(indexName, "_key") {
-		expectedPrefix := tableName + "_"
-		return strings.HasPrefix(indexName, expectedPrefix) && postgresConstraintPattern.MatchString(indexName)
-	}
-
-	// MySQL/MariaDB pattern: simple column name for single-column unique constraints
-	// MySQL automatically creates indexes with the same name as the column for UNIQUE constraints
-	if len(columns) == 1 {
-		// Only consider it constraint-based if the index name matches the column name,
-		// and it does NOT match custom index patterns (e.g., does not start with "idx_" or "index_").
-		// We don't check mysqlTableColumnsPattern here because simple column names like "email"
-		// don't match that pattern (it requires table_column format).
-		return indexName == columns[0] &&
-			!customIndexPattern.MatchString(indexName)
-	}
-
-	// MySQL/MariaDB constraint-based indexes with "uk_" prefix
-	if mysqlUKPattern.MatchString(indexName) {
-		return true
-	}
-
-	// Be more conservative about table_column patterns - only consider it constraint-based
-	// if it follows a very specific pattern and doesn't look like a custom index name
-	if isMySQLConstraintBasedUniqueIndex(indexName, tableName) {
-		return true
-	}
-
-	return false
-}
-
-// Constraints compares constraint definitions between generated and database schemas.
-//
-// This function identifies differences in table-level constraints such as EXCLUDE,
-// CHECK, UNIQUE, PRIMARY KEY, and FOREIGN KEY constraints. It compares constraints
-// defined through Go struct annotations with constraints that exist in the database.
-//
-// # Constraint Types Supported
-//
-//   - EXCLUDE: PostgreSQL EXCLUDE constraints for preventing conflicts
-//   - CHECK: Table-level CHECK constraints for data validation
-//   - UNIQUE: Table-level UNIQUE constraints spanning multiple columns
-//   - PRIMARY KEY: Composite primary key constraints
-//   - FOREIGN KEY: Table-level foreign key constraints
-//
-// # Comparison Logic
-//
-// The function performs constraint comparison by:
-//  1. **Constraint Discovery**: Creates lookup maps for efficient constraint comparison
-//  2. **Addition Detection**: Identifies constraints in generated schema but not in database
-//  3. **Removal Detection**: Identifies constraints in database but not in generated schema
-//
-// # Database Schema Constraints
-//
-// The function currently focuses on constraints defined through schema annotations.
-// Database-introspected constraints are not yet fully supported, so this function
-// primarily detects constraint additions from the generated schema.
-//
-// # Example Usage
-//
-//	// Compare constraints between schemas
-//	compare.Constraints(generated, database, diff)
-//
-//	// Check for constraint changes
-//	if len(diff.ConstraintsAdded) > 0 {
-//		log.Printf("Found %d new constraints to add", len(diff.ConstraintsAdded))
-//	}
-//
-// # Parameters
-//
-//   - generated: Target schema parsed from Go struct annotations
-//   - database: Current database schema from database introspection
-//   - diff: Schema difference structure to populate with constraint changes
-//
-// # Limitations
-//
-//   - Database constraint introspection is not yet fully implemented
-//   - Currently focuses on constraint additions from generated schema
-//   - Constraint modifications are not yet detected
-func Constraints(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff, opts *config.CompareOptions) {
-	dialect := ""
-	if opts != nil {
-		dialect = opts.Dialect
-	}
-
-	// Create maps for detailed constraint comparison
-	genConstraints := make(map[string]goschema.Constraint)
-	for _, constraint := range generated.Constraints {
-		// Use table.constraint_name as the key for comparison to handle constraints with same names in different tables
-		key := constraint.Table + "." + constraint.Name
-		genConstraints[key] = constraint
-	}
-
-	// Synthesize table-level Constraint entries from field-level `check=`
-	// annotations so they participate in drift comparison alongside table
-	// constraints from `//migrator:schema:constraint`. Only synthesized for
-	// columns that already exist in the database — new tables/columns get
-	// their CHECK inline via CREATE TABLE / ALTER TABLE ADD COLUMN, and
-	// double-emitting an ALTER TABLE ADD CONSTRAINT would fail because the
-	// constraint is created in the same migration step.
-	for _, synthesized := range synthesizeFieldLevelCheckConstraints(generated, database) {
-		key := synthesized.Table + "." + synthesized.Name
-		// Don't clobber an explicit table-level constraint that happens to
-		// share the same name.
-		if _, exists := genConstraints[key]; !exists {
-			genConstraints[key] = synthesized
-		}
-	}
-
-	for _, synthesized := range synthesizeTablePrimaryKeyConstraints(generated, database, dialect) {
-		key := synthesized.Table + "." + synthesized.Name
-		// Don't clobber an explicit table-level constraint that happens to
-		// share the same name.
-		if _, exists := genConstraints[key]; !exists {
-			genConstraints[key] = synthesized
-		}
-	}
-
-	// Synthesize table-level Constraint entries from field-level `foreign=`
-	// annotations so on_delete / on_update drift on an existing field-level
-	// FK participates in comparison (issue #189), mirroring the field-level
-	// CHECK synthesis above. Only synthesized for columns that already exist
-	// in the database — new tables/columns get their FK inline via CREATE
-	// TABLE / ALTER TABLE ADD CONSTRAINT, so emitting an ADD CONSTRAINT here
-	// would double-create it in the same migration step.
-	//
-	// synthesizedFKKeys records the table.constraint_name of every synthesized
-	// field-level FK so isFieldLevelConstraint can let the matching DB-side FK
-	// through to the comparison instead of filtering it out — otherwise
-	// foreignKeyConstraintChanged would never run for field-level FKs.
-	synthesizedFKKeys := make(map[string]struct{})
-	for _, synthesized := range synthesizeFieldLevelForeignKeyConstraints(generated, database) {
-		key := synthesized.Table + "." + synthesized.Name
-		synthesizedFKKeys[key] = struct{}{}
-		// Don't clobber an explicit table-level constraint that happens to
-		// share the same name.
-		if _, exists := genConstraints[key]; !exists {
-			genConstraints[key] = synthesized
-		}
-	}
-
-	// Create map of existing database constraints, filtering out field-level constraints
-	dbConstraints := make(map[string]types.DBConstraint)
-	for _, constraint := range database.Constraints {
-		// Skip field-level constraints that are represented in field definitions
-		if isFieldLevelConstraint(constraint, generated, synthesizedFKKeys) {
-			continue
-		}
-
-		// Use table.constraint_name as the key for comparison
-		key := constraint.QualifiedTableName() + "." + constraint.Name
-		dbConstraints[key] = constraint
-	}
-
-	// Find added constraints (constraints in generated schema but not in database)
-	for constraintKey, genConstraint := range genConstraints {
-		if _, exists := dbConstraints[constraintKey]; !exists {
-			diff.ConstraintsAdded = append(diff.ConstraintsAdded, genConstraint.Name)
-			diff.ConstraintsAddedWithTables = appendConstraintAddition(diff.ConstraintsAddedWithTables, genConstraint)
-		}
-	}
-
-	// Find removed constraints (constraints in database but not in generated schema)
-	for constraintKey, dbConstraint := range dbConstraints {
-		if _, exists := genConstraints[constraintKey]; !exists {
-			diff.ConstraintsRemoved = append(diff.ConstraintsRemoved, dbConstraint.Name)
-			diff.ConstraintsRemovedWithTables = appendConstraintRemoval(diff.ConstraintsRemovedWithTables, dbConstraint)
-		}
-	}
-
-	// Find modified constraints (constraints that exist in both but have different definitions)
-	for constraintKey, genConstraint := range genConstraints {
-		if dbConstraint, exists := dbConstraints[constraintKey]; exists {
-			if constraintDefinitionsChanged(genConstraint, dbConstraint, dialect) {
-				// For now, treat modified constraints as removed + added
-				// In the future, we could add a ConstraintsModified field to SchemaDiff
-				diff.ConstraintsRemoved = append(diff.ConstraintsRemoved, dbConstraint.Name)
-				diff.ConstraintsRemovedWithTables = appendConstraintRemoval(diff.ConstraintsRemovedWithTables, dbConstraint)
-				diff.ConstraintsAdded = append(diff.ConstraintsAdded, genConstraint.Name)
-				diff.ConstraintsAddedWithTables = appendConstraintAddition(diff.ConstraintsAddedWithTables, genConstraint)
-			}
-		}
-	}
-
-	// Sort for consistent output. Planners pair the bare name lists with the
-	// *WithTables slices through name-keyed maps, not by index, so each list
-	// can be sorted independently.
-	sort.Strings(diff.ConstraintsAdded)
-	sort.Strings(diff.ConstraintsRemoved)
-	sort.Slice(diff.ConstraintsAddedWithTables, func(i, j int) bool {
-		a, b := diff.ConstraintsAddedWithTables[i], diff.ConstraintsAddedWithTables[j]
-		if a.TableName != b.TableName {
-			return a.TableName < b.TableName
-		}
-		return a.Name < b.Name
-	})
-	sort.Slice(diff.ConstraintsRemovedWithTables, func(i, j int) bool {
-		a, b := diff.ConstraintsRemovedWithTables[i], diff.ConstraintsRemovedWithTables[j]
-		if a.TableName != b.TableName {
-			return a.TableName < b.TableName
-		}
-		return a.Name < b.Name
-	})
-}
-
-// appendConstraintRemoval records the table-qualified removal info for a
-// database constraint that is being dropped (or modified, which is expressed as
-// remove + add). Dialects that need the owning table and a type-specific drop
-// syntax — MySQL/MariaDB FOREIGN KEY uses DROP FOREIGN KEY rather than DROP
-// CONSTRAINT — read this parallel slice instead of the bare name list.
-func appendConstraintRemoval(infos []difftypes.ConstraintRemovalInfo, dbConstraint types.DBConstraint) []difftypes.ConstraintRemovalInfo {
-	return append(infos, difftypes.ConstraintRemovalInfo{
-		Name:      dbConstraint.Name,
-		TableName: dbConstraint.QualifiedTableName(),
-		Type:      dbConstraint.Type,
-	})
-}
-
-// appendConstraintAddition records the table-qualified definition of a
-// constraint that is being added (or modified, which is expressed as remove +
-// add). The bare ConstraintsAdded name list cannot disambiguate a field-level
-// FK whose name repeats across several tables — the canonical case being an
-// embedded inline-relation mixin (e.g. fk_entity_tenant on every table that
-// embeds a tenant-aware base struct, issue #197). Planners read this parallel
-// slice to emit one correctly-targeted ALTER TABLE per real host table instead
-// of re-deriving the table from a field's Go struct name (which, for a mixin,
-// is not a table at all). For a unique-named constraint this carries exactly
-// one entry and matches the legacy field-scan behavior.
-func appendConstraintAddition(infos []difftypes.ConstraintAdditionInfo, genConstraint goschema.Constraint) []difftypes.ConstraintAdditionInfo {
-	return append(infos, difftypes.ConstraintAdditionInfo{
-		Name:            genConstraint.Name,
-		TableName:       genConstraint.Table,
-		Type:            genConstraint.Type,
-		Columns:         append([]string(nil), genConstraint.Columns...),
-		NullsDistinct:   cloneBoolPtr(genConstraint.NullsDistinct),
-		CheckExpression: genConstraint.CheckExpression,
-		ForeignTable:    genConstraint.ForeignTable,
-		ForeignColumn:   genConstraint.ForeignColumn,
-		ForeignColumns:  append([]string(nil), genConstraint.ForeignColumnsOrDefault()...),
-		OnDelete:        genConstraint.OnDelete,
-		OnUpdate:        genConstraint.OnUpdate,
-	})
-}
-
-// constraintDefinitionsChanged compares constraint definitions between generated and database schemas
-// to detect if a constraint needs to be recreated due to definition changes.
-func constraintDefinitionsChanged(genConstraint goschema.Constraint, dbConstraint types.DBConstraint, dialect string) bool {
-	// Basic constraint type comparison
-	if genConstraint.Type != dbConstraint.Type {
-		return true
-	}
-
-	// Type-specific comparisons
-	switch genConstraint.Type {
-	case "EXCLUDE":
-		return excludeConstraintChanged(genConstraint, dbConstraint)
-	case "CHECK":
-		return checkConstraintChanged(genConstraint, dbConstraint)
-	case "UNIQUE":
-		return uniqueConstraintChanged(genConstraint, dbConstraint)
-	case "PRIMARY KEY":
-		return primaryKeyConstraintChanged(genConstraint, dbConstraint)
-	case "FOREIGN KEY":
-		return foreignKeyConstraintChanged(genConstraint, dbConstraint, dialect)
-	default:
-		// For unknown constraint types, assume no change
-		return false
-	}
-}
-
-func primaryKeyConstraintChanged(genConstraint goschema.Constraint, dbConstraint types.DBConstraint) bool {
-	return !slices.Equal(genConstraint.Columns, dbConstraint.ColumnNamesOrDefault())
-}
-
-// excludeConstraintChanged compares EXCLUDE constraint definitions
-func excludeConstraintChanged(genConstraint goschema.Constraint, dbConstraint types.DBConstraint) bool {
-	// Compare using method
-	if genConstraint.UsingMethod != getStringValue(dbConstraint.UsingMethod) {
-		return true
-	}
-
-	// Compare exclude elements
-	if genConstraint.ExcludeElements != getStringValue(dbConstraint.ExcludeElements) {
-		return true
-	}
-
-	// Compare WHERE condition
-	if genConstraint.WhereCondition != getStringValue(dbConstraint.WhereCondition) {
-		return true
-	}
-
-	return false
-}
-
-// checkConstraintChanged compares CHECK constraint definitions.
-func checkConstraintChanged(genConstraint goschema.Constraint, dbConstraint types.DBConstraint) bool {
-	dbClause := getStringValue(dbConstraint.CheckClause)
-	if strings.TrimSpace(genConstraint.CheckExpression) == "" || strings.TrimSpace(dbClause) == "" {
-		return false
-	}
-	if checkExpressionHasUnsupportedRewrite(genConstraint.CheckExpression, dbClause) {
-		return false
-	}
-	return normalizeCheckExpression(genConstraint.CheckExpression) != normalizeCheckExpression(dbClause)
-}
-
-// uniqueConstraintChanged compares UNIQUE constraint definitions
-func uniqueConstraintChanged(genConstraint goschema.Constraint, dbConstraint types.DBConstraint) bool {
-	return !stringSetsEqual(genConstraint.Columns, dbConstraint.ColumnNamesOrDefault()) ||
-		!boolPtrEqual(genConstraint.NullsDistinct, dbConstraint.NullsDistinct)
-}
-
-func normalizeCheckExpression(expr string) string {
-	expr = trimBalancedCheckParens(strings.TrimSpace(expr))
-
-	var b strings.Builder
-	inString := false
-	for i := 0; i < len(expr); i++ {
-		ch := expr[i]
-		if ch == '\'' {
-			b.WriteByte(ch)
-			if inString && i+1 < len(expr) && expr[i+1] == '\'' {
-				i++
-				b.WriteByte('\'')
-				continue
-			}
-			inString = !inString
-			continue
-		}
-		if !inString && ch == ':' && i+1 < len(expr) && expr[i+1] == ':' && checkCastFollowsLiteral(expr, i) {
-			i += 2
-			for i < len(expr) && isCheckCastChar(rune(expr[i])) {
-				i++
-			}
-			i--
-			continue
-		}
-		if !inString && unicode.IsSpace(rune(ch)) {
-			continue
-		}
-		if !inString && ch == '`' {
-			continue
-		}
-		if !inString && ch >= 'A' && ch <= 'Z' {
-			ch += 'a' - 'A'
-		}
-		b.WriteByte(ch)
-	}
-	return b.String()
-}
-
-func trimBalancedCheckParens(expr string) string {
-	for {
-		trimmed := strings.TrimSpace(expr)
-		if len(trimmed) < 2 || trimmed[0] != '(' || trimmed[len(trimmed)-1] != ')' {
-			return trimmed
-		}
-		if !checkOuterParensWrap(trimmed) {
-			return trimmed
-		}
-		expr = trimmed[1 : len(trimmed)-1]
-	}
-}
-
-func checkOuterParensWrap(expr string) bool {
-	depth := 0
-	inString := false
-	for i := 0; i < len(expr); i++ {
-		ch := expr[i]
-		if ch == '\'' {
-			if inString && i+1 < len(expr) && expr[i+1] == '\'' {
-				i++
-				continue
-			}
-			inString = !inString
-			continue
-		}
-		if inString {
-			continue
-		}
-
-		switch ch {
-		case '(':
-			depth++
-		case ')':
-			depth--
-			if depth < 0 || (depth == 0 && i != len(expr)-1) {
-				return false
-			}
-		}
-	}
-	return depth == 0
-}
-
-func isCheckCastChar(ch rune) bool {
-	return unicode.IsLetter(ch) || unicode.IsDigit(ch) || ch == '_' || ch == '[' || ch == ']' || ch == '.'
-}
-
-func checkCastFollowsLiteral(expr string, castIndex int) bool {
-	for i := castIndex - 1; i >= 0; i-- {
-		ch := expr[i]
-		if unicode.IsSpace(rune(ch)) {
-			continue
-		}
-		return ch == '\'' || (ch >= '0' && ch <= '9')
-	}
-	return false
-}
-
-func checkExpressionHasUnsupportedRewrite(generated, database string) bool {
-	db := strings.ToLower(database)
-	return checkExpressionContainsInOperator(generated) && strings.Contains(db, "any") && strings.Contains(db, "array[")
-}
-
-func checkExpressionContainsInOperator(expr string) bool {
-	inString := false
-	for i := 0; i < len(expr)-1; i++ {
-		ch := expr[i]
-		if ch == '\'' {
-			if inString && i+1 < len(expr) && expr[i+1] == '\'' {
-				i++
-				continue
-			}
-			inString = !inString
-			continue
-		}
-		if inString || (ch != 'i' && ch != 'I') || (expr[i+1] != 'n' && expr[i+1] != 'N') {
-			continue
-		}
-		prevIdent := i > 0 && isCheckIdentChar(rune(expr[i-1]))
-		nextIdent := i+2 < len(expr) && isCheckIdentChar(rune(expr[i+2]))
-		if !prevIdent && !nextIdent {
-			return true
-		}
-	}
-	return false
-}
-
-func isCheckIdentChar(ch rune) bool {
-	return unicode.IsLetter(ch) || unicode.IsDigit(ch) || ch == '_'
-}
-
-func stringSetsEqual(left, right []string) bool {
-	left = uniqueStringsPreserveOrder(left)
-	right = uniqueStringsPreserveOrder(right)
-	sort.Strings(left)
-	sort.Strings(right)
-	return slices.Equal(left, right)
-}
-
-func boolPtrEqual(left, right *bool) bool {
-	if left == nil || right == nil {
-		return left == right
-	}
-	return *left == *right
-}
-
-func cloneBoolPtr(value *bool) *bool {
-	if value == nil {
-		return nil
-	}
-	clone := *value
-	return &clone
-}
-
-// foreignKeyConstraintChanged compares FOREIGN KEY constraint definitions.
-//
-// Referential actions are normalized before comparison (see
-// normalizeReferentialAction): databases report the default action as
-// "NO ACTION" via information_schema.referential_constraints, whereas a Go
-// annotation that omits on_delete/on_update leaves the value empty. Without
-// the normalization the two would never match for FKs declared without an
-// explicit action, producing a perpetual drop+add loop on every `generate`
-// (the same hazard checkConstraintChanged guards against for CHECK clauses).
-func foreignKeyConstraintChanged(genConstraint goschema.Constraint, dbConstraint types.DBConstraint, dialect string) bool {
-	// Compare local columns.
-	if !slices.Equal(genConstraint.Columns, uniqueStringsPreserveOrder(dbConstraint.ColumnNamesOrDefault())) {
-		return true
-	}
-
-	// Compare referenced table
-	if !foreignTableRefMatches(genConstraint.ForeignTable, dbConstraint) {
-		return true
-	}
-
-	// Compare referenced columns.
-	if !slices.Equal(genConstraint.ForeignColumnsOrDefault(), uniqueStringsPreserveOrder(dbConstraint.ForeignColumnsOrDefault())) {
-		return true
-	}
-
-	// Compare delete rule
-	if normalizeReferentialAction(genConstraint.OnDelete, dialect) != normalizeReferentialAction(getStringValue(dbConstraint.DeleteRule), dialect) {
-		return true
-	}
-
-	// Compare update rule
-	if normalizeReferentialAction(genConstraint.OnUpdate, dialect) != normalizeReferentialAction(getStringValue(dbConstraint.UpdateRule), dialect) {
-		return true
-	}
-
-	return false
-}
-
-func uniqueStringsPreserveOrder(values []string) []string {
-	result := make([]string, 0, len(values))
-	seen := make(map[string]struct{}, len(values))
-	for _, value := range values {
-		if _, ok := seen[value]; ok {
-			continue
-		}
-		seen[value] = struct{}{}
-		result = append(result, value)
-	}
-	return result
-}
-
-func foreignTableRefMatches(generated string, dbConstraint types.DBConstraint) bool {
-	generated = strings.TrimSpace(generated)
-	if generated == "" {
-		return dbConstraint.ForeignTable == nil
-	}
-	if strings.Contains(generated, ".") {
-		return generated == dbConstraint.QualifiedForeignTableName()
-	}
-	return generated == getStringValue(dbConstraint.ForeignTable)
-}
-
-// normalizeReferentialAction canonicalizes an ON DELETE / ON UPDATE action so
-// that semantically identical values compare equal across the generated and
-// introspected sides.
-//
-// SQL treats an omitted referential action as NO ACTION. PostgreSQL, MySQL and
-// MariaDB all report the default through
-// information_schema.referential_constraints, while a Go field annotation that
-// simply omits on_delete/on_update yields an empty string. Trimming,
-// upper-casing, and folding "" into "NO ACTION" makes those equivalent and
-// keeps an unchanged FK a no-op on repeated runs.
-//
-// Dialect-specific RESTRICT handling: MariaDB reports an unspecified action as
-// RESTRICT (PostgreSQL and MySQL report NO ACTION), and InnoDB treats RESTRICT
-// and NO ACTION identically. For the MySQL family RESTRICT is therefore folded
-// to NO ACTION so an unchanged FK does not loop drop+add forever. PostgreSQL
-// distinguishes RESTRICT (checked immediately) from NO ACTION (deferrable) at
-// DDL level, so the fold is NOT applied there — doing so would mask a genuine
-// RESTRICT <-> NO ACTION change the user intended.
-func normalizeReferentialAction(action, dialect string) string {
-	normalized := strings.ToUpper(strings.TrimSpace(action))
-	if normalized == "" {
-		return "NO ACTION"
-	}
-	if normalized == "RESTRICT" && isMySQLFamily(dialect) {
-		return "NO ACTION"
-	}
-	return normalized
-}
-
-// isMySQLFamily reports whether the dialect is MySQL or MariaDB, which share the
-// InnoDB referential-action semantics (RESTRICT == NO ACTION).
-func isMySQLFamily(dialect string) bool {
-	switch strings.ToLower(strings.TrimSpace(dialect)) {
-	case "mysql", "mariadb":
-		return true
-	default:
-		return false
-	}
-}
-
-// getStringValue safely extracts string value from a pointer, returning empty string if nil
-func getStringValue(ptr *string) string {
-	if ptr == nil {
-		return ""
-	}
-	return *ptr
-}
-
-// isFieldLevelConstraint determines if a database constraint represents a field-level constraint
-// that is already represented in the field definitions (NOT NULL, PRIMARY KEY, UNIQUE, FOREIGN KEY).
-//
-// synthesizedFKKeys holds the table.constraint_name of every field-level FK that
-// Constraints() synthesized into the generated set (see
-// synthesizeFieldLevelForeignKeyConstraints). When a DB-side FK has a synthesized
-// counterpart it is NOT treated as field-level here, so it stays in the
-// comparison and on_delete / on_update drift flows through
-// foreignKeyConstraintChanged (issue #189). FKs without a synthesized
-// counterpart (e.g. a column that is not yet in the database, which never gets
-// synthesized) keep the previous filter-out behavior.
-func isFieldLevelConstraint(dbConstraint types.DBConstraint, generated *goschema.Database, synthesizedFKKeys map[string]struct{}) bool {
-	// Create a map of table.column -> field for quick lookup
-	fieldMap := make(map[string]goschema.Field)
-	for _, field := range generated.Fields {
-		// Get table name for this field
-		tableName := field.StructName // default to struct name
-		for _, table := range generated.Tables {
-			if table.StructName == field.StructName {
-				tableName = table.QualifiedName()
-				break
-			}
-		}
-		key := tableName + "." + field.Name
-		fieldMap[key] = field
-	}
-
-	// Check if this constraint corresponds to a field-level constraint
-	switch dbConstraint.Type {
-	case "NOT NULL":
-		// Check if there's a field with not_null=true for this column
-		key := dbConstraint.QualifiedTableName() + "." + getConstraintColumn(dbConstraint)
-		if field, exists := fieldMap[key]; exists && !field.Nullable {
-			return true
-		}
-	case "PRIMARY KEY":
-		// Check if there's a field with primary=true for this column
-		key := dbConstraint.QualifiedTableName() + "." + getConstraintColumn(dbConstraint)
-		if field, exists := fieldMap[key]; exists && field.Primary {
-			return true
-		}
-	case "UNIQUE":
-		// Check if there's a field with unique=true for this column
-		key := dbConstraint.QualifiedTableName() + "." + getConstraintColumn(dbConstraint)
-		if field, exists := fieldMap[key]; exists && field.Unique {
-			return true
-		}
-	case "FOREIGN KEY":
-		// A field-level FK that was synthesized into the generated set (see
-		// synthesizeFieldLevelForeignKeyConstraints) must participate in the
-		// comparison so on_delete / on_update drift is detected (issue #189).
-		// Letting the DB-side FK through means it gets matched by name against
-		// the synthesized entry, so add / remove / action-change cases all
-		// flow through the standard Constraints() comparison path. Match on
-		// the constraint name rather than the column, because the synthesized
-		// name is keyed on table.constraint_name and getConstraintColumn does
-		// not always resolve the FK column.
-		if _, synthesized := synthesizedFKKeys[dbConstraint.QualifiedTableName()+"."+dbConstraint.Name]; synthesized {
-			return false
-		}
-		// Check if there's a field with foreign key reference for this column.
-		// No synthesized counterpart (e.g. the column is not yet in the
-		// database): keep the historical behavior and treat it as field-level
-		// so it is owned by the column/table lifecycle.
-		key := dbConstraint.QualifiedTableName() + "." + getConstraintColumn(dbConstraint)
-		if field, exists := fieldMap[key]; exists && field.Foreign != "" {
-			return true
-		}
-	case "CHECK":
-		// PostgreSQL exposes NOT NULL declarations as synthetic CHECK rows in
-		// information_schema with a `<table>_<column>_not_null` naming
-		// convention (and, in PG18, as named NOT NULL constraints in
-		// pg_constraint). Ptah never emits these directly — they are owned by
-		// the column's NOT NULL clause, which the table/column lifecycle
-		// already handles:
-		//   - field exists in target with NOT NULL → covered by the CREATE
-		//     TABLE / ALTER COLUMN SET NOT NULL emitted elsewhere.
-		//   - field exists in target as nullable → an ALTER COLUMN DROP
-		//     NOT NULL handles it.
-		//   - field has been dropped → the column drop cascades.
-		//   - table has been dropped → the table drop cascades.
-		// In every case an explicit ALTER TABLE … DROP CONSTRAINT for the
-		// synthetic name would be redundant at best and illegal at worst (PG
-		// rejects with 42P16 "column X is in a primary key" when the column
-		// is part of a PK). Treat these as field-level always; skip the diff.
-		if strings.Contains(dbConstraint.Name, "_not_null") {
-			return true
-		}
-		// Regular CHECK constraints from `check=` annotations are surfaced
-		// to the diff via synthesized goschema.Constraint entries (see
-		// synthesizeFieldLevelCheckConstraints in Constraints). Letting the
-		// DB constraint participate here means it gets matched against the
-		// synthesized entry by name, so add/remove/expression-change cases
-		// all flow through the standard Constraints() comparison path.
-	}
-
-	return false
-}
-
-// synthesizeFieldLevelCheckConstraints turns each field-level `check=`
-// annotation on an existing database column into a synthetic
-// goschema.Constraint of type CHECK so the standard Constraints() diff path
-// can compare it against the introspected CHECK from pg_constraint.
-//
-// The constraint name follows the user-provided `check_name=` value when set,
-// otherwise it falls back to the PostgreSQL convention
-// "<table>_<column>_check" — which is what PostgreSQL itself uses for
-// unnamed inline column-level CHECKs, so the name lines up with whatever the
-// reader sees on the DB side.
-//
-// Columns that do not yet exist in the database are deliberately skipped:
-// those CHECKs ship inline as part of CREATE TABLE / ALTER TABLE ADD COLUMN,
-// and emitting an ADD CONSTRAINT alongside would attempt to create the same
-// constraint twice in the same migration.
-//
-// Precedence: an explicit table-level constraint declared via
-// `//migrator:schema:constraint` that happens to share the synthesized
-// name wins — synthesis never clobbers it. See the guard in
-// Constraints() where genConstraints is populated.
-func synthesizeFieldLevelCheckConstraints(generated *goschema.Database, database *types.DBSchema) []goschema.Constraint {
-	if generated == nil || database == nil {
-		return nil
-	}
-
-	structToTable := make(map[string]goschema.Table, len(generated.Tables))
-	for _, t := range generated.Tables {
-		structToTable[t.StructName] = t
-	}
-
-	dbColumns := make(map[string]struct{}, 16)
-	for _, t := range database.Tables {
-		for _, c := range t.Columns {
-			dbColumns[t.QualifiedName()+"."+c.Name] = struct{}{}
-		}
-	}
-
-	var synthesized []goschema.Constraint
-	for _, f := range generated.Fields {
-		if f.Check == "" {
-			continue
-		}
-		table, ok := structToTable[f.StructName]
-		tableName := table.QualifiedName()
-		tableLeafName := table.Name
-		if !ok || tableName == "" {
-			tableName = f.StructName
-			tableLeafName = f.StructName
-		}
-		if _, exists := dbColumns[tableName+"."+f.Name]; !exists {
-			continue
-		}
-		name := f.CheckName
-		if name == "" {
-			name = tableLeafName + "_" + f.Name + "_check"
-		}
-		synthesized = append(synthesized, goschema.Constraint{
-			StructName:      f.StructName,
-			Name:            name,
-			Type:            "CHECK",
-			Table:           tableName,
-			CheckExpression: f.Check,
-		})
-	}
-	return synthesized
-}
-
-func synthesizeTablePrimaryKeyConstraints(
-	generated *goschema.Database,
-	database *types.DBSchema,
-	dialect string,
-) []goschema.Constraint {
-	if generated == nil || database == nil {
-		return nil
-	}
-
-	dbTables := make(map[string]struct{}, len(database.Tables))
-	for _, table := range database.Tables {
-		dbTables[table.QualifiedName()] = struct{}{}
-	}
-
-	var synthesized []goschema.Constraint
-	for _, table := range generated.Tables {
-		columns := tablePrimaryKeyColumns(table)
-		if len(columns) == 0 {
-			continue
-		}
-		if _, exists := dbTables[table.QualifiedName()]; !exists {
-			continue
-		}
-
-		name := tablePrimaryKeyConstraintName(table, database.Constraints, dialect)
-		synthesized = append(synthesized, goschema.Constraint{
-			StructName: table.StructName,
-			Name:       name,
-			Type:       "PRIMARY KEY",
-			Table:      table.QualifiedName(),
-			Columns:    append([]string(nil), columns...),
-		})
-	}
-	return synthesized
-}
-
-func tablePrimaryKeyConstraintName(table goschema.Table, dbConstraints []types.DBConstraint, dialect string) string {
-	for _, constraint := range dbConstraints {
-		if constraint.Type == "PRIMARY KEY" && constraint.QualifiedTableName() == table.QualifiedName() {
-			return constraint.Name
-		}
-	}
-
-	if isMySQLFamily(dialect) {
-		return "PRIMARY"
-	}
-	return table.Name + "_pkey"
-}
-
-// synthesizeFieldLevelForeignKeyConstraints turns each field-level `foreign=`
-// annotation on an existing database column into a synthetic
-// goschema.Constraint of type FOREIGN KEY so the standard Constraints() diff
-// path can compare it against the introspected FK from
-// information_schema.referential_constraints. This is what makes on_delete /
-// on_update drift on a pre-existing field-level FK observable (issue #189).
-//
-// The constraint name follows the user-provided `foreign_key_name=` value when
-// set, otherwise it falls back to the conventional generated name from
-// fromschema.GenerateForeignKeyName ("fk_<table>_<column>"), which is the name
-// the planner emits when it creates the FK, so the synthesized name lines up
-// with whatever the reader sees on the DB side.
-//
-// Columns that do not yet exist in the database are deliberately skipped: those
-// FKs ship inline as part of CREATE TABLE / ALTER TABLE ADD COLUMN, and emitting
-// an ADD CONSTRAINT alongside would attempt to create the same constraint twice
-// in the same migration. This mirrors synthesizeFieldLevelCheckConstraints
-// exactly and is what keeps added-table generation untouched.
-//
-// Precedence: an explicit table-level constraint declared via
-// `//migrator:schema:constraint` that happens to share the synthesized name
-// wins — synthesis never clobbers it (see the guard in Constraints()).
-func synthesizeFieldLevelForeignKeyConstraints(generated *goschema.Database, database *types.DBSchema) []goschema.Constraint {
-	if generated == nil || database == nil {
-		return nil
-	}
-
-	dbColumns := make(map[string]struct{}, 16)
-	for _, t := range database.Tables {
-		for _, c := range t.Columns {
-			dbColumns[t.QualifiedName()+"."+c.Name] = struct{}{}
-		}
-	}
-
-	// Iterate the fields that actually materialize on each concrete table,
-	// not the raw parse result. A `foreign=` annotation declared on an
-	// embedded inline-relation mixin (e.g. a TenantGroupAwareEntityID base
-	// struct carrying tenant_id/group_id/created_by_user_id FKs) lives on the
-	// mixin's StructName, which is NOT a table. Synthesizing against
-	// f.StructName therefore produced a constraint targeting the Go struct
-	// name (ALTER TABLE TenantGroupAwareEntityID ...) once per embedding host,
-	// all collapsed onto the same bogus name (issue #197). Resolving via the
-	// same CREATE-path embedded expansion that TableColumns uses gives one
-	// field per real host table with the host's StructName, so each embedding
-	// table gets its own correctly-targeted FK.
-	//
-	// dedupe guards against a host that both declares a field directly and
-	// inherits one of the same (table, constraint name) from a mixin.
-	dedupe := make(map[string]struct{})
-	var synthesized []goschema.Constraint
-	for _, f := range resolveTableFields(generated) {
-		if f.Foreign == "" {
-			continue
-		}
-		// resolveTableFields only returns fields that belong to a real table,
-		// tagged with that table's name. An empty tableName would mean the
-		// field is not part of any table, so skip it rather than synthesize
-		// against a struct name.
-		tableName := f.qualifiedTableName
-		if tableName == "" {
-			continue
-		}
-		if _, exists := dbColumns[tableName+"."+f.Name]; !exists {
-			continue
-		}
-		name := f.ForeignKeyName
-		if name == "" {
-			name = fromschema.GenerateForeignKeyName(f.tableName, f.Name)
-		}
-		// Reuse the canonical generate-path parser so the synthesized table /
-		// column always match exactly what the planner emits (issue #189
-		// follow-up: a single source of truth removes the latent two-parser
-		// divergence). A malformed foreign= reference yields nil and is skipped
-		// rather than synthesizing a garbage constraint.
-		fkRef := fromschema.ParseForeignKeyReference(f.Foreign)
-		if fkRef == nil {
-			continue
-		}
-		dedupeKey := tableName + "." + name
-		if _, seen := dedupe[dedupeKey]; seen {
-			continue
-		}
-		dedupe[dedupeKey] = struct{}{}
-		synthesized = append(synthesized, goschema.Constraint{
-			StructName:     f.StructName,
-			Name:           name,
-			Type:           "FOREIGN KEY",
-			Table:          tableName,
-			Columns:        []string{f.Name},
-			ForeignTable:   fkRef.Table,
-			ForeignColumn:  fkRef.Column,
-			ForeignColumns: fkRef.ReferencedColumns(),
-			OnDelete:       f.OnDelete,
-			OnUpdate:       f.OnUpdate,
-		})
-	}
-	return synthesized
-}
-
-// resolvedField is a goschema.Field paired with the concrete database table it
-// materializes on. Fields declared directly on a table struct carry that
-// table's name; fields contributed by an embedded inline / inline-relation
-// mixin are expanded once per embedding host and carry the host table's name.
-type resolvedField struct {
-	goschema.Field
-	tableName          string
-	qualifiedTableName string
-}
-
-// resolveTableFields expands every table's field set the same way the CREATE
-// and column-diff paths do (processEmbeddedFieldsForStruct), tagging each
-// resulting field with its concrete host table name. This is the single source
-// of truth for "which fields end up as columns on which real table", so any
-// field-level synthesis (FK drift, and any future field-level constraint
-// synthesis) targets host tables rather than mixin struct names (issue #197).
-//
-// Only fields whose owning struct is a declared table are returned: a
-// `foreign=` annotation on a mixin that is never embedded, or on a struct that
-// is not a //migrator:schema:table, has no concrete table and must not be
-// synthesized.
-func resolveTableFields(generated *goschema.Database) []resolvedField {
-	if generated == nil {
-		return nil
-	}
-
-	var resolved []resolvedField
-	for _, table := range generated.Tables {
-		// Direct fields declared on the table struct itself.
-		for _, f := range generated.Fields {
-			if f.StructName == table.StructName {
-				resolved = append(resolved, resolvedField{Field: f, tableName: table.Name, qualifiedTableName: table.QualifiedName()})
-			}
-		}
-		// Fields contributed by embedded mixins (inline + inline-relation),
-		// each already rewritten to the host struct name by the expansion.
-		for _, f := range processEmbeddedFieldsForStruct(generated.EmbeddedFields, generated.Fields, table.StructName) {
-			resolved = append(resolved, resolvedField{Field: f, tableName: table.Name, qualifiedTableName: table.QualifiedName()})
-		}
-	}
-	return resolved
-}
-
-// getConstraintColumn extracts the column name from a constraint
-// This is a simplified implementation - in practice, constraints can span multiple columns
-func getConstraintColumn(constraint types.DBConstraint) string {
-	// For single-column constraints, try to extract column name from constraint name
-	// This is database-specific and may need refinement
-
-	// Use the ColumnName field if available (MySQL/MariaDB provide this directly)
-	if constraint.ColumnName != "" {
-		return constraint.ColumnName
-	}
-
-	// PostgreSQL NOT NULL constraints often follow pattern: schema_table_column_not_null
-	if constraint.Type == "NOT NULL" && strings.Contains(constraint.Name, "_not_null") {
-		return extractPostgreSQLNotNullColumn(constraint)
-	}
-
-	// For PRIMARY KEY constraints: table_pkey (PostgreSQL) or PRIMARY (MySQL/MariaDB)
-	if constraint.Type == "PRIMARY KEY" {
-		if strings.HasSuffix(constraint.Name, "_pkey") {
-			// PostgreSQL pattern: table_pkey
-			// This is a table-level primary key, we need to check if it's single-column
-			// For now, assume single-column primary keys are field-level
-			return "id" // common convention, but this is a limitation
-		} else if constraint.Name == "PRIMARY" {
-			// MySQL/MariaDB pattern: PRIMARY
-			// For single-column primary keys, assume it's the id field
-			return "id" // common convention, but this is a limitation
-		}
-	}
-
-	// For UNIQUE constraints: table_column_key (PostgreSQL) or column name (MySQL/MariaDB)
-	if constraint.Type == "UNIQUE" {
-		if strings.HasSuffix(constraint.Name, "_key") {
-			// PostgreSQL pattern
-			return extractPostgreSQLUniqueColumn(constraint)
-		}
-
-		// MySQL/MariaDB often use the column name as the constraint name for single-column unique constraints
-		return constraint.Name
-	}
-
-	// For other constraints, return empty string (will not match any field)
-	return ""
-}
-
-// extractPostgreSQLNotNullColumn extracts column name from PostgreSQL NOT NULL constraint
-func extractPostgreSQLNotNullColumn(constraint types.DBConstraint) string {
-	parts := strings.Split(constraint.Name, "_")
-	if len(parts) < 3 {
-		return ""
-	}
-
-	// Remove schema prefix (2200_) and suffix (_not_null)
-	if !strings.HasPrefix(constraint.Name, "2200_") {
-		return ""
-	}
-
-	// Format: 2200_table_column_not_null -> extract column
-	tableParts := strings.Split(constraint.Name[5:], "_not_null")
-	if len(tableParts) == 0 {
-		return ""
-	}
-
-	remaining := tableParts[0]
-	// Remove table name prefix to get column
-	tablePrefix := constraint.TableName + "_"
-	if strings.HasPrefix(remaining, tablePrefix) {
-		return remaining[len(tablePrefix):]
-	}
-
-	return ""
-}
-
-// extractPostgreSQLUniqueColumn extracts column name from PostgreSQL UNIQUE constraint
-func extractPostgreSQLUniqueColumn(constraint types.DBConstraint) string {
-	parts := strings.Split(constraint.Name, "_")
-	if len(parts) < 3 {
-		return ""
-	}
-
-	// Remove table prefix and _key suffix
-	tablePrefix := constraint.TableName + "_"
-	if !strings.HasPrefix(constraint.Name, tablePrefix) {
-		return ""
-	}
-
-	remaining := constraint.Name[len(tablePrefix):]
-	if !strings.HasSuffix(remaining, "_key") {
-		return ""
-	}
-
-	return remaining[:len(remaining)-4] // remove "_key"
-}
-
-// isMySQLConstraintBasedUniqueIndex checks if an index follows MySQL/MariaDB constraint-based patterns.
-// This helper function encapsulates the complex logic for detecting MySQL/MariaDB constraint-based
-// unique indexes that follow table_column naming patterns but are not custom indexes.
-func isMySQLConstraintBasedUniqueIndex(indexName, tableName string) bool {
-	return mysqlTableColumnsPattern.MatchString(indexName) &&
-		strings.HasPrefix(indexName, tableName+"_") &&
-		!customIndexPattern.MatchString(indexName)
-}
-
-// Indexes performs index comparison between generated and database schemas with intelligent filtering.
-//
-// This function handles the comparison of database indexes, which requires careful
-// filtering to avoid false positives from automatically generated indexes (primary
-// keys, unique constraints) that are managed by the database system rather than
-// explicitly defined in the schema.
-//
-// # Index Filtering Logic
-//
-// The function applies intelligent filtering to focus on user-defined indexes:
-//
-// **Generated Schema Indexes**:
-//   - Includes all explicitly defined indexes from Go struct annotations
-//   - These are indexes the developer intentionally created for performance
-//
-// **Database Schema Indexes**:
-//   - Excludes primary key indexes (automatically created with PRIMARY KEY constraints)
-//   - Excludes constraint-based unique indexes (automatically created with UNIQUE constraints)
-//   - Includes explicitly defined unique indexes (created via schema annotations)
-//   - Includes manually created performance indexes
-//
-// This filtering prevents false positives where the system would suggest removing
-// automatically generated constraint indexes that are essential for constraint enforcement,
-// while still allowing comparison of explicitly defined unique indexes.
-//
-// # Example Scenarios
-//
-// **Performance index addition**:
-//
-//	```go
-//	type User struct {
-//	    Email string `db:"email" index:"idx_users_email"`
-//	}
-//	```
-//	- Generated schema defines "idx_users_email"
-//	- Database doesn't have this index
-//	- Result: "idx_users_email" added to diff.IndexesAdded
-//
-// **Unused index removal**:
-//   - Database has "idx_old_search" index
-//   - Generated schema doesn't define this index
-//   - Result: "idx_old_search" added to diff.IndexesRemoved
-//
-// **Automatic index filtering**:
-//   - Database has "users_pkey" (primary key index) - filtered out
-//   - Database has "users_email_key" (constraint-based unique index) - filtered out
-//   - Database has "users_tenant_email_idx" (explicitly defined unique index) - included for comparison
-//
-// # Algorithm Details
-//
-// 1. **Set Creation**: Converts index lists to boolean maps for O(1) lookup
-// 2. **Filtering**: Applies database-side filtering for automatic indexes
-// 3. **Comparison**: Performs set difference operations to find additions/removals
-// 4. **Sorting**: Ensures deterministic output for consistent migrations
-//
-// # Parameters
-//
-//   - generated: Target schema parsed from Go struct annotations
-//   - database: Current database schema from executor introspection
-//   - diff: SchemaDiff structure to populate with discovered differences
-//
-// # Side Effects
-//
-// Modifies the provided diff parameter by populating:
-//   - diff.IndexesAdded: Indexes that need to be created
-//   - diff.IndexesRemoved: User-defined indexes that can be safely removed
-//
-// # Safety Considerations
-//
-// Index operations are generally safe:
-//   - Adding indexes improves performance but doesn't affect data
-//   - Removing indexes may impact query performance but doesn't cause data loss
-//   - Primary key and unique constraint indexes are protected from removal
-//
-// # Performance Impact
-//
-// - Time Complexity: O(n + m) where n=generated indexes, m=database indexes
-// - Space Complexity: O(n + m) for the boolean maps
-// - Index operations can be expensive on large tables in production
-func Indexes(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff) {
-	// Create sets for comparison
-	genIndexes := make(map[string]goschema.Index)
-	for _, index := range generated.Indexes {
-		genIndexes[index.Name] = index
-	}
-
-	// MySQL/MariaDB transparently create a backing index for every FOREIGN KEY,
-	// named after the constraint, when no usable index already exists on the
-	// referencing column. That index is owned by the FK, not by the user's
-	// schema, so it must not be reported as a removable index — otherwise an
-	// unchanged field-level FK round-trips to a spurious DROP INDEX (issue #189
-	// follow-up). PostgreSQL does not auto-create FK indexes, so this set is
-	// naturally empty there and the filter is a no-op. Keyed by table.index so a
-	// constraint name only suppresses the index on its own table.
-	fkBackedIndexes := make(map[string]struct{}, len(database.Constraints))
-	uniqueConstraintIndexes := make(map[string]struct{}, len(database.Constraints))
-	for _, c := range database.Constraints {
-		switch c.Type {
-		case "FOREIGN KEY":
-			fkBackedIndexes[c.QualifiedTableName()+"."+c.Name] = struct{}{}
-		case "UNIQUE":
-			uniqueConstraintIndexes[c.QualifiedTableName()+"."+c.Name] = struct{}{}
-		}
-	}
-
-	dbIndexes := make(map[string]types.DBIndex)
-	for _, index := range database.Indexes {
-		// Skip primary key indexes as they're handled with tables
-		if index.IsPrimary {
-			continue
-		}
-
-		// Skip constraint-based unique indexes (automatically created by UNIQUE constraints)
-		// but allow explicitly defined unique indexes (created via schema annotations)
-		if index.IsUnique && isConstraintBasedUniqueIndex(index.Name, index.TableName, index.Columns) {
-			continue
-		}
-		if _, ok := uniqueConstraintIndexes[index.QualifiedTableName()+"."+index.Name]; ok {
-			continue
-		}
-
-		// Skip MySQL/MariaDB FK-backing indexes (named after the FK constraint
-		// on the same table). They are auto-managed by the foreign key.
-		if _, ok := fkBackedIndexes[index.QualifiedTableName()+"."+index.Name]; ok {
-			continue
-		}
-
-		dbIndexes[index.Name] = index
-	}
-
-	// Find added and removed indexes
-	for indexName, genIndex := range genIndexes {
-		dbIndex, exists := dbIndexes[indexName]
-		switch {
-		case !exists:
-			diff.IndexesAdded = append(diff.IndexesAdded, indexName)
-		case indexDefinitionsChanged(genIndex, dbIndex):
-			diff.IndexesAdded = append(diff.IndexesAdded, indexName)
-			diff.IndexesRemoved = append(diff.IndexesRemoved, indexName)
-			diff.IndexesRemovedWithTables = append(diff.IndexesRemovedWithTables, difftypes.IndexRemovalInfo{
-				Name:      indexName,
-				TableName: dbIndex.QualifiedTableName(),
-			})
-		}
-	}
-
-	for indexName, dbIndex := range dbIndexes {
-		if _, exists := genIndexes[indexName]; !exists {
-			diff.IndexesRemoved = append(diff.IndexesRemoved, indexName)
-			diff.IndexesRemovedWithTables = append(diff.IndexesRemovedWithTables, difftypes.IndexRemovalInfo{
-				Name:      indexName,
-				TableName: dbIndex.QualifiedTableName(),
-			})
-		}
-	}
-
-	// Sort for consistent output
-	sort.Strings(diff.IndexesAdded)
-	sort.Strings(diff.IndexesRemoved)
-
-	// Sort the detailed removal info by index name
-	sort.Slice(diff.IndexesRemovedWithTables, func(i, j int) bool {
-		return diff.IndexesRemovedWithTables[i].Name < diff.IndexesRemovedWithTables[j].Name
-	})
-}
-
-func indexDefinitionsChanged(genIndex goschema.Index, dbIndex types.DBIndex) bool {
-	return !boolPtrEqual(genIndex.NullsDistinct, dbIndex.NullsDistinct) ||
-		indexPredicateChanged(genIndex.Condition, dbIndex.Condition)
-}
-
-func indexPredicateChanged(generated, database string) bool {
-	if strings.TrimSpace(generated) == "" || strings.TrimSpace(database) == "" {
-		return strings.TrimSpace(generated) != strings.TrimSpace(database)
-	}
-	if checkExpressionHasUnsupportedRewrite(generated, database) {
-		return false
-	}
-	return normalizePredicate(generated) != normalizePredicate(database)
-}
-
-func normalizePredicate(value string) string {
-	return normalizeCheckExpression(value)
-}
-
-// compareNamedItems is a generic helper function that compares two maps of named items
-// and returns the names of items that are added (in generated but not in database)
-// and removed (in database but not in generated).
-//
-// This helper eliminates code duplication between Functions and RLSPolicies comparison logic.
-func compareNamedItems[T, U any](generated map[string]T, database map[string]U) (added, removed []string) {
-	// Find added items (in generated but not in database)
-	for name := range generated {
-		if _, exists := database[name]; !exists {
-			added = append(added, name)
-		}
-	}
-
-	// Find removed items (in database but not in generated)
-	for name := range database {
-		if _, exists := generated[name]; !exists {
-			removed = append(removed, name)
-		}
-	}
-
-	return added, removed
-}
-
-// Functions performs PostgreSQL function comparison between generated and database schemas.
-//
-// This function handles the comparison of PostgreSQL custom functions, which are
-// PostgreSQL-specific features used for stored procedures, triggers, and custom
-// business logic. Functions are compared by name and their complete definition.
-//
-// # Function Comparison Logic
-//
-// **Generated Schema Functions**:
-//   - Includes all functions defined in Go struct annotations
-//   - These are functions the developer intentionally created
-//
-// **Database Schema Functions**:
-//   - Includes all user-defined functions from the database
-//   - Excludes system functions and built-in PostgreSQL functions
-//   - Excludes extension-owned functions (filtered by database reader)
-//
-// # Extension Function Filtering
-//
-// Extension-owned functions are automatically excluded by the database reader to prevent
-// migration issues. Extension functions cannot be dropped independently and attempting
-// to do so will cause migration failures. Common extensions with functions include:
-//   - btree_gin: Functions like gin_btree_consistent, gin_extract_*
-//   - pg_trgm: Functions like similarity, word_similarity, gin_trgm_*
-//
-// # Function Modification Detection
-//
-// Functions are considered modified if any of the following differ:
-//   - Parameters (type, names, order)
-//   - Return type
-//   - Function body/implementation
-//   - Language (plpgsql, sql, etc.)
-//   - Security context (DEFINER vs INVOKER)
-//   - Volatility (STABLE, IMMUTABLE, VOLATILE)
-//
-// # Example Scenarios
-//
-// **Function addition**:
-//   - Generated schema defines "get_current_tenant_id()"
-//   - Database doesn't have this function
-//   - Result: "get_current_tenant_id" added to diff.FunctionsAdded
-//
-// **Function removal**:
-//   - Database has "old_helper_function()"
-//   - Generated schema doesn't define this function
-//   - Result: "old_helper_function" added to diff.FunctionsRemoved
-//
-// **Function modification**:
-//   - Both have "calculate_total()" function
-//   - Generated: different body or parameters
-//   - Result: FunctionDiff added to diff.FunctionsModified
-//
-// # Parameters
-//
-//   - generated: Target schema parsed from Go struct annotations
-//   - database: Current database schema from database introspection
-//   - diff: SchemaDiff structure to populate with discovered differences
-//
-// # Side Effects
-//
-// Modifies the provided diff parameter by populating:
-//   - diff.FunctionsAdded: Functions that need to be created
-//   - diff.FunctionsRemoved: Functions that exist in database but not in target schema
-//   - diff.FunctionsModified: Functions with definition differences
-//
-// # Output Consistency
-//
-// Results are sorted alphabetically for consistent output across multiple runs.
-func Functions(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff) {
-	// Build lookup maps for function comparison
-	generatedFunctionMap := make(map[string]goschema.Function)
-	for _, fn := range generated.Functions {
-		generatedFunctionMap[fn.Name] = fn
-	}
-
-	databaseFunctionMap := make(map[string]types.DBFunction)
-	for _, fn := range database.Functions {
-		databaseFunctionMap[fn.Name] = fn
-	}
-
-	// Use generic comparison helper for add/remove detection
-	addedFunctions, removedFunctions := compareNamedItems(generatedFunctionMap, databaseFunctionMap)
-	diff.FunctionsAdded = append(diff.FunctionsAdded, addedFunctions...)
-	diff.FunctionsRemoved = append(diff.FunctionsRemoved, removedFunctions...)
-
-	// Detect function definition modifications
-	for functionName, generatedFunction := range generatedFunctionMap {
-		if databaseFunction, functionExists := databaseFunctionMap[functionName]; functionExists {
-			functionComparison := FunctionDefinitions(generatedFunction, databaseFunction)
-			if len(functionComparison.Changes) > 0 {
-				diff.FunctionsModified = append(diff.FunctionsModified, functionComparison)
-			}
-		}
-	}
-
-	// Ensure consistent ordering of results
-	sort.Strings(diff.FunctionsAdded)
-	sort.Strings(diff.FunctionsRemoved)
-	sort.Slice(diff.FunctionsModified, func(i, j int) bool {
-		return diff.FunctionsModified[i].FunctionName < diff.FunctionsModified[j].FunctionName
-	})
-}
-
-// Views compares view definitions between generated and database schemas.
-func Views(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff) {
-	generatedViews := make(map[string]goschema.View)
-	for _, view := range generated.Views {
-		generatedViews[view.Name] = view
-	}
-
-	databaseViews := make(map[string]types.DBView)
-	for _, view := range database.Views {
-		databaseViews[view.QualifiedName()] = view
-	}
-
-	addedViews, removedViews := compareNamedItems(generatedViews, databaseViews)
-	diff.ViewsAdded = append(diff.ViewsAdded, addedViews...)
-	diff.ViewsRemoved = append(diff.ViewsRemoved, removedViews...)
-
-	for viewName, generatedView := range generatedViews {
-		if databaseView, exists := databaseViews[viewName]; exists {
-			viewDiff := ViewDefinitions(generatedView, databaseView)
-			if len(viewDiff.Changes) > 0 {
-				diff.ViewsModified = append(diff.ViewsModified, viewDiff)
-			}
-		}
-	}
-
-	sort.Strings(diff.ViewsAdded)
-	sort.Strings(diff.ViewsRemoved)
-	sort.Slice(diff.ViewsModified, func(i, j int) bool {
-		return diff.ViewsModified[i].ViewName < diff.ViewsModified[j].ViewName
-	})
-}
-
-// MaterializedViews compares materialized view definitions between generated
-// and database schemas.
-func MaterializedViews(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff) {
-	generatedViews := make(map[string]goschema.MaterializedView)
-	for _, view := range generated.MaterializedViews {
-		view.Canonicalize()
-		generatedViews[view.Name] = view
-	}
-
-	databaseViews := make(map[string]types.DBMatView)
-	for _, view := range database.MatViews {
-		databaseViews[view.QualifiedName()] = view
-	}
-
-	addedViews, removedViews := compareNamedItems(generatedViews, databaseViews)
-	diff.MaterializedViewsAdded = append(diff.MaterializedViewsAdded, addedViews...)
-	diff.MaterializedViewsRemoved = append(diff.MaterializedViewsRemoved, removedViews...)
-
-	for viewName, generatedView := range generatedViews {
-		if databaseView, exists := databaseViews[viewName]; exists {
-			viewDiff := MaterializedViewDefinitions(generatedView, databaseView)
-			if len(viewDiff.Changes) > 0 {
-				diff.MaterializedViewsModified = append(diff.MaterializedViewsModified, viewDiff)
-			}
-		}
-	}
-
-	sort.Strings(diff.MaterializedViewsAdded)
-	sort.Strings(diff.MaterializedViewsRemoved)
-	sort.Slice(diff.MaterializedViewsModified, func(i, j int) bool {
-		return diff.MaterializedViewsModified[i].ViewName < diff.MaterializedViewsModified[j].ViewName
-	})
-}
-
-// Triggers compares trigger definitions between generated and database schemas.
-func Triggers(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff) {
-	generatedTriggers := make(map[string]goschema.Trigger)
-	for _, trigger := range generated.Triggers {
-		trigger.Canonicalize()
-		generatedTriggers[triggerKey(trigger.Table, trigger.Name)] = trigger
-	}
-
-	databaseTriggers := make(map[string]types.DBTrigger)
-	for _, trigger := range database.Triggers {
-		databaseTriggers[triggerKey(trigger.QualifiedTable(), trigger.Name)] = trigger
-	}
-
-	for key, trigger := range generatedTriggers {
-		if _, exists := databaseTriggers[key]; !exists {
-			diff.TriggersAdded = append(diff.TriggersAdded, difftypes.TriggerRef{
-				TriggerName: trigger.Name,
-				TableName:   trigger.Table,
-			})
-		}
-	}
-	for key, trigger := range databaseTriggers {
-		if _, exists := generatedTriggers[key]; !exists {
-			diff.TriggersRemoved = append(diff.TriggersRemoved, difftypes.TriggerRef{
-				TriggerName: trigger.Name,
-				TableName:   trigger.QualifiedTable(),
-			})
-		}
-	}
-	for key, generatedTrigger := range generatedTriggers {
-		if databaseTrigger, exists := databaseTriggers[key]; exists {
-			triggerDiff := TriggerDefinitions(generatedTrigger, databaseTrigger)
-			if len(triggerDiff.Changes) > 0 {
-				diff.TriggersModified = append(diff.TriggersModified, triggerDiff)
-			}
-		}
-	}
-
-	sortTriggerRefs(diff.TriggersAdded)
-	sortTriggerRefs(diff.TriggersRemoved)
-	sort.Slice(diff.TriggersModified, func(i, j int) bool {
-		if diff.TriggersModified[i].TableName == diff.TriggersModified[j].TableName {
-			return diff.TriggersModified[i].TriggerName < diff.TriggersModified[j].TriggerName
-		}
-		return diff.TriggersModified[i].TableName < diff.TriggersModified[j].TableName
-	})
-}
-
-func triggerKey(tableName, triggerName string) string {
-	return tableName + "." + triggerName
-}
-
-func sortTriggerRefs(refs []difftypes.TriggerRef) {
-	sort.Slice(refs, func(i, j int) bool {
-		if refs[i].TableName == refs[j].TableName {
-			return refs[i].TriggerName < refs[j].TriggerName
-		}
-		return refs[i].TableName < refs[j].TableName
-	})
-}
-
-// RLSPolicies performs PostgreSQL RLS policy comparison between generated and database schemas.
-//
-// This function handles the comparison of Row-Level Security policies, which are
-// PostgreSQL-specific security features used for multi-tenant data isolation and
-// fine-grained access control. Policies are compared by name and their complete definition.
-//
-// # RLS Policy Comparison Logic
-//
-// **Generated Schema Policies**:
-//   - Includes all RLS policies defined in Go struct annotations
-//   - These are policies the developer intentionally created for data security
-//
-// **Database Schema Policies**:
-//   - Includes all user-defined RLS policies from the database
-//   - Excludes system-generated policies (if any)
-//
-// # Policy Modification Detection
-//
-// Policies are considered modified if any of the following differ:
-//   - Policy type (FOR clause: ALL, SELECT, INSERT, UPDATE, DELETE)
-//   - Target roles (TO clause)
-//   - USING expression for row filtering
-//   - WITH CHECK expression for INSERT/UPDATE validation
-//
-// # Example Scenarios
-//
-// **Policy addition**:
-//   - Generated schema defines "user_tenant_isolation" policy
-//   - Database doesn't have this policy
-//   - Result: "user_tenant_isolation" added to diff.RLSPoliciesAdded
-//
-// **Policy removal**:
-//   - Database has "old_security_policy" policy
-//   - Generated schema doesn't define this policy
-//   - Result: "old_security_policy" added to diff.RLSPoliciesRemoved
-//
-// **Policy modification**:
-//   - Both have "tenant_isolation" policy
-//   - Generated: different USING expression or target roles
-//   - Result: RLSPolicyDiff added to diff.RLSPoliciesModified
-//
-// # Parameters
-//
-//   - generated: Target schema parsed from Go struct annotations
-//   - database: Current database schema from database introspection
-//   - diff: SchemaDiff structure to populate with discovered differences
-//
-// # Side Effects
-//
-// Modifies the provided diff parameter by populating:
-//   - diff.RLSPoliciesAdded: Policies that need to be created
-//   - diff.RLSPoliciesRemoved: Policies that exist in database but not in target schema
-//   - diff.RLSPoliciesModified: Policies with definition differences
-//
-// # Output Consistency
-//
-// Results are sorted alphabetically for consistent output across multiple runs.
-func RLSPolicies(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff) {
-	// Build lookup maps for RLS policy comparison
-	generatedPolicyMap := make(map[string]goschema.RLSPolicy)
-	for _, rlsPolicy := range generated.RLSPolicies {
-		generatedPolicyMap[rlsPolicy.Name] = rlsPolicy
-	}
-
-	databasePolicyMap := make(map[string]types.DBRLSPolicy)
-	for _, rlsPolicy := range database.RLSPolicies {
-		databasePolicyMap[rlsPolicy.Name] = rlsPolicy
-	}
-
-	// Find added policies (inline logic to avoid duplication detection)
-	for policyName := range generatedPolicyMap {
-		if _, exists := databasePolicyMap[policyName]; !exists {
-			diff.RLSPoliciesAdded = append(diff.RLSPoliciesAdded, policyName)
-		}
-	}
-
-	// Find removed policies
-	for policyName, dbPolicy := range databasePolicyMap {
-		if _, exists := generatedPolicyMap[policyName]; !exists {
-			policyRef := difftypes.RLSPolicyRef{
-				PolicyName: policyName,
-				TableName:  dbPolicy.Table,
-			}
-			diff.RLSPoliciesRemoved = append(diff.RLSPoliciesRemoved, policyRef)
-		}
-	}
-
-	// Detect policy definition modifications
-	for policyName, generatedPolicy := range generatedPolicyMap {
-		if databasePolicy, policyExists := databasePolicyMap[policyName]; policyExists {
-			policyComparison := RLSPolicyDefinitions(generatedPolicy, databasePolicy)
-			if len(policyComparison.Changes) > 0 {
-				diff.RLSPoliciesModified = append(diff.RLSPoliciesModified, policyComparison)
-			}
-		}
-	}
-
-	// Ensure consistent ordering of results
-	sort.Strings(diff.RLSPoliciesAdded)
-	sort.Slice(diff.RLSPoliciesRemoved, func(i, j int) bool {
-		return diff.RLSPoliciesRemoved[i].PolicyName < diff.RLSPoliciesRemoved[j].PolicyName
-	})
-	sort.Slice(diff.RLSPoliciesModified, func(i, j int) bool {
-		return diff.RLSPoliciesModified[i].PolicyName < diff.RLSPoliciesModified[j].PolicyName
-	})
-}
-
-// RLSEnabledTables performs RLS enablement comparison between generated and database schemas.
-//
-// This function handles the comparison of RLS enablement status on tables, determining
-// which tables need RLS enabled or disabled based on the target schema definition.
-//
-// # RLS Enablement Logic
-//
-// **Generated Schema RLS Tables**:
-//   - Includes all tables that should have RLS enabled according to annotations
-//   - These are tables the developer wants to secure with row-level policies
-//
-// **Database Schema RLS Tables**:
-//   - Includes all tables that currently have RLS enabled in the database
-//   - Determined by checking pg_class.relrowsecurity for PostgreSQL
-//
-// # Example Scenarios
-//
-// **RLS enablement**:
-//   - Generated schema specifies RLS should be enabled on "users" table
-//   - Database doesn't have RLS enabled on "users"
-//   - Result: "users" added to diff.RLSEnabledTablesAdded
-//
-// **RLS disablement**:
-//   - Database has RLS enabled on "legacy_table"
-//   - Generated schema doesn't specify RLS for "legacy_table"
-//   - Result: "legacy_table" added to diff.RLSEnabledTablesRemoved
-//
-// # Parameters
-//
-//   - generated: Target schema parsed from Go struct annotations
-//   - database: Current database schema from database introspection
-//   - diff: SchemaDiff structure to populate with discovered differences
-//
-// # Side Effects
-//
-// Modifies the provided diff parameter by populating:
-//   - diff.RLSEnabledTablesAdded: Tables that need RLS enabled
-//   - diff.RLSEnabledTablesRemoved: Tables that need RLS disabled
-//
-// # Output Consistency
-//
-// Results are sorted alphabetically for consistent output across multiple runs.
-func RLSEnabledTables(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff) {
-	// Create sets for comparison
-	genRLSTables := make(map[string]bool)
-	for _, rlsTable := range generated.RLSEnabledTables {
-		genRLSTables[rlsTable.Table] = true
-	}
-
-	dbRLSTables := make(map[string]bool)
-	for _, table := range database.Tables {
-		if table.RLSEnabled {
-			dbRLSTables[table.QualifiedName()] = true
-		}
-	}
-
-	// Find tables that need RLS enabled
-	for tableName := range genRLSTables {
-		if !dbRLSTables[tableName] {
-			diff.RLSEnabledTablesAdded = append(diff.RLSEnabledTablesAdded, tableName)
-		}
-	}
-
-	// Find tables that need RLS disabled
-	for tableName := range dbRLSTables {
-		if !genRLSTables[tableName] {
-			diff.RLSEnabledTablesRemoved = append(diff.RLSEnabledTablesRemoved, tableName)
-		}
-	}
-
-	// Sort for consistent output
-	sort.Strings(diff.RLSEnabledTablesAdded)
-	sort.Strings(diff.RLSEnabledTablesRemoved)
-}
-
-// Roles performs PostgreSQL role comparison between generated and database schemas.
-//
-// This function handles the comparison of PostgreSQL database roles, which are
-// used for authentication, authorization, and access control. Roles are compared
-// by name and their complete attribute definition.
-//
-// # Role Comparison Logic
-//
-// **Generated Schema Roles**:
-//   - Includes all roles defined in Go struct annotations
-//   - These are roles the developer intentionally created for application security
-//
-// **Database Schema Roles**:
-//   - Includes all user-defined roles from the database
-//   - Excludes system roles (pg_*, postgres) for safety
-//
-// # Role Modification Detection
-//
-// Roles are considered modified if any of the following differ:
-//   - Login capability (can the role login)
-//   - Password (encrypted password hash)
-//   - Superuser status (administrative privileges)
-//   - CreateDB capability (can create databases)
-//   - CreateRole capability (can create other roles)
-//   - Inherit capability (inherits privileges from granted roles)
-//   - Replication capability (can initiate replication)
-//
-// # Example Scenarios
-//
-// **Role addition**:
-//   - Generated schema defines "app_user" role
-//   - Database doesn't have this role
-//   - Result: "app_user" added to diff.RolesAdded
-//
-// **Role removal**:
-//   - Roles are NOT automatically marked for removal for safety reasons
-//   - Existing roles not defined in schema are left untouched
-//   - Manual role removal should be done by DBAs when needed
-//
-// **Role modification**:
-//   - Both have "api_user" role
-//   - Generated: different login capability or privileges
-//   - Result: RoleDiff added to diff.RolesModified
-//
-// # Parameters
-//
-//   - generated: Target schema parsed from Go struct annotations
-//   - database: Current database schema from database introspection
-//   - diff: SchemaDiff structure to populate with discovered differences
-//
-// # Side Effects
-//
-// Modifies the provided diff parameter by populating:
-//   - diff.RolesAdded: Roles that need to be created
-//   - diff.RolesRemoved: Always empty (roles are not automatically removed for safety)
-//   - diff.RolesModified: Roles with attribute differences
-//
-// # Output Consistency
-//
-// Results are sorted alphabetically for consistent output across multiple runs.
-func Roles(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff) {
-	// Build lookup maps for role comparison
-	generatedRoleMap := make(map[string]goschema.Role)
-	for _, role := range generated.Roles {
-		generatedRoleMap[role.Name] = role
-	}
-
-	databaseRoleMap := make(map[string]types.DBRole)
-	for _, role := range database.Roles {
-		databaseRoleMap[role.Name] = role
-	}
-
-	// Find added roles
-	for roleName := range generatedRoleMap {
-		if _, exists := databaseRoleMap[roleName]; !exists {
-			diff.RolesAdded = append(diff.RolesAdded, roleName)
-		}
-	}
-
-	// Note: We intentionally do not automatically mark roles for removal.
-	// Roles are security-sensitive objects that may be created by DBAs,
-	// other applications, or infrastructure setup. Automatic removal could
-	// be dangerous and break authentication/authorization.
-	// If role removal is needed, it should be done explicitly by the DBA.
-
-	// Detect role attribute modifications
-	for roleName, generatedRole := range generatedRoleMap {
-		if databaseRole, roleExists := databaseRoleMap[roleName]; roleExists {
-			roleComparison := RoleDefinitions(generatedRole, databaseRole)
-			if len(roleComparison.Changes) > 0 {
-				diff.RolesModified = append(diff.RolesModified, roleComparison)
-			}
-		}
-	}
-
-	// Ensure consistent ordering of results
-	sort.Strings(diff.RolesAdded)
-	sort.Strings(diff.RolesRemoved)
-	sort.Slice(diff.RolesModified, func(i, j int) bool {
-		return diff.RolesModified[i].RoleName < diff.RolesModified[j].RoleName
-	})
-}
-
-// Grants compares PostgreSQL role privilege grants.
-func Grants(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff) {
-	generatedGrantMap := make(map[string]difftypes.GrantRef)
-	generatedGrantRoles := make(map[string]bool)
-	for _, grant := range generated.Grants {
-		generatedGrantRoles[grant.Role] = true
-		for _, ref := range grantRefsFromGenerated(grant) {
-			generatedGrantMap[grantRefIdentityKey(ref)] = ref
-		}
-	}
-
-	managedRoles := make(map[string]bool)
-	for _, role := range generated.Roles {
-		managedRoles[role.Name] = true
-	}
-
-	databaseGrantMapForAdditions := make(map[string]difftypes.GrantRef)
-	databaseGrantMapForRemovals := make(map[string]difftypes.GrantRef)
-	for _, grant := range database.Grants {
-		ref := grantRefFromDatabase(grant)
-		key := grantRefIdentityKey(ref)
-		if managedRoles[ref.Role] || generatedGrantRoles[ref.Role] {
-			databaseGrantMapForAdditions[key] = ref
-		}
-		if managedRoles[ref.Role] {
-			databaseGrantMapForRemovals[key] = ref
-		}
-	}
-
-	for key, ref := range generatedGrantMap {
-		databaseRef, exists := databaseGrantMapForAdditions[key]
-		if !exists {
-			diff.GrantsAdded = append(diff.GrantsAdded, ref)
-			continue
-		}
-		if ref.WithOption && !databaseRef.WithOption {
-			diff.GrantOptionsAdded = append(diff.GrantOptionsAdded, ref)
-		}
-		if !ref.WithOption && databaseRef.WithOption && managedRoles[ref.Role] {
-			diff.GrantOptionsRevoked = append(diff.GrantOptionsRevoked, databaseRef)
-		}
-	}
-	for key, ref := range databaseGrantMapForRemovals {
-		if _, exists := generatedGrantMap[key]; !exists {
-			diff.GrantsRemoved = append(diff.GrantsRemoved, ref)
-		}
-	}
-
-	sortGrantRefs(diff.GrantsAdded)
-	sortGrantRefs(diff.GrantsRemoved)
-	sortGrantRefs(diff.GrantOptionsAdded)
-	sortGrantRefs(diff.GrantOptionsRevoked)
-}
-
-func grantRefsFromGenerated(grant goschema.Grant) []difftypes.GrantRef {
-	grant.Canonicalize()
-	objectType := "TABLE"
-	objectName := grant.OnTable
-	if grant.OnSchema != "" {
-		objectType = "SCHEMA"
-		objectName = grant.OnSchema
-	}
-	refs := make([]difftypes.GrantRef, 0, len(grant.Privileges))
-	for _, privilege := range grant.Privileges {
-		refs = append(refs, difftypes.GrantRef{
-			Role:       grant.Role,
-			Privilege:  strings.ToUpper(strings.TrimSpace(privilege)),
-			ObjectType: objectType,
-			ObjectName: objectName,
-			WithOption: grant.WithOption,
-		})
-	}
-	return refs
-}
-
-func grantRefFromDatabase(grant types.DBGrant) difftypes.GrantRef {
-	objectType := strings.ToUpper(strings.TrimSpace(grant.ObjectType))
-	objectName := grant.QualifiedTarget()
-	if objectType == "SCHEMA" {
-		objectName = grant.ObjectName
-	}
-	return difftypes.GrantRef{
-		Role:       strings.TrimSpace(grant.Role),
-		Privilege:  strings.ToUpper(strings.TrimSpace(grant.Privilege)),
-		ObjectType: objectType,
-		ObjectName: objectName,
-		WithOption: grant.WithOption,
-	}
-}
-
-func grantRefIdentityKey(ref difftypes.GrantRef) string {
-	return strings.Join([]string{
-		strings.TrimSpace(ref.Role),
-		strings.ToUpper(strings.TrimSpace(ref.ObjectType)),
-		strings.TrimSpace(ref.ObjectName),
-		strings.ToUpper(strings.TrimSpace(ref.Privilege)),
-	}, "\x00")
-}
-
-func sortGrantRefs(refs []difftypes.GrantRef) {
-	sort.Slice(refs, func(i, j int) bool {
-		if refs[i].ObjectType != refs[j].ObjectType {
-			return refs[i].ObjectType < refs[j].ObjectType
-		}
-		if refs[i].ObjectName != refs[j].ObjectName {
-			return refs[i].ObjectName < refs[j].ObjectName
-		}
-		if refs[i].Role != refs[j].Role {
-			return refs[i].Role < refs[j].Role
-		}
-		if refs[i].Privilege != refs[j].Privilege {
-			return refs[i].Privilege < refs[j].Privilege
-		}
-		return !refs[i].WithOption && refs[j].WithOption
-	})
-}
-
-// RoleDefinitions compares individual role definitions and returns detailed differences.
-//
-// This function performs attribute-by-attribute comparison of PostgreSQL role definitions,
-// identifying specific changes needed to bring the database role in line with the target
-// role definition. It handles all PostgreSQL role attributes including privileges and capabilities.
-//
-// # Comparison Attributes
-//
-// The function compares the following role attributes:
-//   - **Login**: Whether the role can login to the database
-//   - **Password**: Role password (note: actual passwords are not compared for security)
-//   - **Superuser**: Whether the role has superuser privileges
-//   - **CreateDB**: Whether the role can create databases
-//   - **CreateRole**: Whether the role can create other roles
-//   - **Inherit**: Whether the role inherits privileges from granted roles
-//   - **Replication**: Whether the role can initiate streaming replication
-//
-// # Password Handling
-//
-// Password comparison is handled specially:
-//   - If target role has a password and database role doesn't, it's marked as changed
-//   - If target role has no password and database role has one, no change is recorded
-//   - Actual password values are not compared for security reasons
-//
-// # Change Format
-//
-// Changes are recorded in "old_value -> new_value" format for clarity:
-//   - Boolean attributes: "false -> true" or "true -> false"
-//   - Password: "no_password -> password_set" or similar safe representation
-//
-// # Parameters
-//
-//   - generated: Target role definition from Go struct annotations
-//   - database: Current role definition from database introspection
-//
-// # Return Value
-//
-// Returns a RoleDiff structure containing:
-//   - RoleName: Name of the role being compared
-//   - Changes: Map of attribute changes in "old -> new" format
-//
-// # Example Output
-//
-//	RoleDiff{
-//		RoleName: "app_user",
-//		Changes: map[string]string{
-//			"login": "false -> true",
-//			"createdb": "false -> true",
-//			"password": "no_password -> password_set",
-//		},
-//	}
-func RoleDefinitions(generated goschema.Role, database types.DBRole) difftypes.RoleDiff {
-	roleDiff := difftypes.RoleDiff{
-		RoleName: generated.Name,
-		Changes:  make(map[string]string),
-	}
-
-	// Compare login capability
-	if generated.Login != database.Login {
-		roleDiff.Changes["login"] = fmt.Sprintf("%t -> %t", database.Login, generated.Login)
-	}
-
-	// Compare password (special handling for security)
-	// We only detect if a password needs to be set, not compare actual values
-	if generated.Password != "" && !database.HasPassword {
-		// If target has password but database role doesn't, mark for update
-		roleDiff.Changes["password"] = "password_update_required"
-	}
-
-	// Compare superuser status
-	if generated.Superuser != database.Superuser {
-		roleDiff.Changes["superuser"] = fmt.Sprintf("%t -> %t", database.Superuser, generated.Superuser)
-	}
-
-	// Compare createdb capability
-	if generated.CreateDB != database.CreateDB {
-		roleDiff.Changes["createdb"] = fmt.Sprintf("%t -> %t", database.CreateDB, generated.CreateDB)
-	}
-
-	// Compare createrole capability
-	if generated.CreateRole != database.CreateRole {
-		roleDiff.Changes["createrole"] = fmt.Sprintf("%t -> %t", database.CreateRole, generated.CreateRole)
-	}
-
-	// Compare inherit capability
-	if generated.Inherit != database.Inherit {
-		roleDiff.Changes["inherit"] = fmt.Sprintf("%t -> %t", database.Inherit, generated.Inherit)
-	}
-
-	// Compare replication capability
-	if generated.Replication != database.Replication {
-		roleDiff.Changes["replication"] = fmt.Sprintf("%t -> %t", database.Replication, generated.Replication)
-	}
-
-	return roleDiff
-}
-
-// processEmbeddedFieldsForStruct processes embedded fields for a specific struct and generates corresponding schema fields.
-//
-// This function implements the core logic for transforming embedded fields into database schema fields
-// according to their specified embedding mode. It processes only embedded fields that belong to the
-// specified structName.
-//
-// This is a local implementation to replace the obsolete transform package.
-func processEmbeddedFieldsForStruct(embeddedFields []goschema.EmbeddedField, allFields []goschema.Field, structName string) []goschema.Field {
-	var generatedFields []goschema.Field
-
-	// Process each embedded field definition
-	for _, embedded := range embeddedFields {
-		// Filter: only process embedded fields for the target struct
-		if embedded.StructName != structName {
-			continue
-		}
-
-		switch embedded.Mode {
-		case "inline":
-			// INLINE MODE: Expand embedded struct fields as individual table columns (with recursive support)
-			generatedFields = processEmbeddedInlineModeRecursiveForSchemaDiff(generatedFields, embedded, allFields, embeddedFields, structName)
-		case "json":
-			// JSON MODE: Create a single JSON/JSONB column for the embedded struct
-			jsonField := goschema.Field{
-				StructName: structName,
-				FieldName:  embedded.EmbeddedTypeName,
-				Name:       embedded.Name,
-				Type:       embedded.Type,
-				Nullable:   embedded.Nullable,
-				Comment:    embedded.Comment,
-			}
-			generatedFields = append(generatedFields, jsonField)
-		case "relation":
-			// RELATION MODE: Create a foreign key field
-			// Create platform-specific overrides for MySQL/MariaDB compatibility
-			overrides := make(map[string]map[string]string)
-			overrides["mysql"] = map[string]string{"type": "INT"}
-			overrides["mariadb"] = map[string]string{"type": "INT"}
-
-			relationField := goschema.Field{
-				StructName: structName,
-				FieldName:  embedded.EmbeddedTypeName + "ID",
-				Name:       embedded.Field,
-				Type:       "INTEGER", // Default to INTEGER for foreign keys
-				Foreign:    embedded.Ref,
-				OnDelete:   embedded.OnDelete, // Mirror of fromschema.processEmbeddedRelationMode — keeps the diff path in agreement with the generate path on FK actions (#117).
-				OnUpdate:   embedded.OnUpdate,
-				Comment:    embedded.Comment,
-				Overrides:  overrides, // Platform-specific type overrides
-			}
-			generatedFields = append(generatedFields, relationField)
-		case "skip":
-			// SKIP MODE: Do nothing - completely ignore this embedded field
-			continue
-		default:
-			// DEFAULT MODE: Fall back to inline behavior for unrecognized modes (with recursive support)
-			generatedFields = processEmbeddedInlineModeRecursiveForSchemaDiff(generatedFields, embedded, allFields, embeddedFields, structName)
-		}
-	}
-
-	return generatedFields
-}
-
-// processEmbeddedInlineModeRecursiveForSchemaDiff recursively processes embedded fields in inline mode for schema diff.
-// This handles nested embedded structs by recursively expanding embedded fields within embedded types.
-func processEmbeddedInlineModeRecursiveForSchemaDiff(generatedFields []goschema.Field, embedded goschema.EmbeddedField, allFields []goschema.Field, allEmbeddedFields []goschema.EmbeddedField, structName string) []goschema.Field {
-	// Step 1: Add direct fields from the embedded type
-	for _, field := range allFields {
-		if field.StructName != embedded.EmbeddedTypeName {
-			continue
-		}
-		// Clone the field and reassign to target struct
-		newField := field
-		newField.StructName = structName
-
-		// Apply prefix to column name if specified
-		if embedded.Prefix != "" {
-			newField.Name = embedded.Prefix + field.Name
-		}
-
-		generatedFields = append(generatedFields, newField)
-	}
-
-	// Step 2: Recursively process embedded fields within the embedded type
-	for _, nestedEmbedded := range allEmbeddedFields {
-		if nestedEmbedded.StructName != embedded.EmbeddedTypeName {
-			continue
-		}
-
-		// Only process inline mode embedded fields recursively
-		if nestedEmbedded.Mode == "inline" {
-			// Create a new embedded field with the target struct name and combined prefix
-			recursiveEmbedded := nestedEmbedded
-			recursiveEmbedded.StructName = structName
-
-			// Combine prefixes: if the parent has a prefix, prepend it to the nested prefix
-			if embedded.Prefix != "" {
-				if recursiveEmbedded.Prefix != "" {
-					recursiveEmbedded.Prefix = embedded.Prefix + recursiveEmbedded.Prefix
-				} else {
-					recursiveEmbedded.Prefix = embedded.Prefix
-				}
-			}
-
-			// Recursively process the nested embedded field
-			generatedFields = processEmbeddedInlineModeRecursiveForSchemaDiff(generatedFields, recursiveEmbedded, allFields, allEmbeddedFields, structName)
-		}
-	}
-
-	return generatedFields
-}
-
-// Extensions performs comprehensive extension comparison between generated and database schemas.
-//
-// This function compares PostgreSQL extensions defined in the target schema (from Go struct annotations)
-// with extensions currently installed in the database. It identifies which extensions need to be
-// added or removed to bring the database in line with the target schema.
-//
-// # Extension Ignore Functionality
-//
-// The function supports ignoring specific extensions through the opts parameter:
-//   - Ignored extensions are filtered out before comparison
-//   - Ignored extensions will never be marked for removal
-//   - Ignored extensions can still be created if defined in the target schema
-//   - If opts is nil, default options are used (ignores "plpgsql")
-//
-// # Comparison Process
-//
-// The function performs comparison in three phases:
-//  1. **Extension Filtering**: Removes ignored extensions from consideration
-//  2. **Extension Discovery**: Creates lookup maps for efficient extension comparison
-//  3. **Extension Diff Analysis**: Identifies added and removed extensions
-//
-// # PostgreSQL Extension Considerations
-//
-// Extensions in PostgreSQL are database-wide objects that provide additional functionality:
-//   - **pg_trgm**: Trigram similarity search and GIN operator classes
-//   - **btree_gin**: GIN indexes for btree-compatible data types
-//   - **postgis**: Geographic data types and functions
-//   - **uuid-ossp**: UUID generation functions
-//   - **plpgsql**: Procedural language (usually pre-installed, commonly ignored)
-//
-// # Extension Detection
-//
-// The function now fully supports extension detection from the database schema, enabling
-// accurate comparison between target and current state. This allows for proper extension
-// lifecycle management including both addition and removal operations.
-//
-// # Parameters
-//
-//   - generated: Target schema parsed from Go struct annotations
-//   - database: Current database schema from executor introspection (includes extensions)
-//   - diff: SchemaDiff structure to populate with discovered differences
-//   - opts: Configuration options for comparison (can be nil for defaults)
-//
-// # Side Effects
-//
-// Modifies the provided diff parameter by populating:
-//   - diff.ExtensionsAdded: Extensions that need to be created
-//   - diff.ExtensionsRemoved: Extensions that exist in database but not in target schema
-//
-// # Example Usage
-//
-//	// Extensions defined in Go annotations
-//	//migrator:schema:extension name="pg_trgm" if_not_exists="true"
-//	//migrator:schema:extension name="btree_gin" if_not_exists="true"
-//	type DatabaseExtensions struct{}
-//
-//	// Database has pg_trgm installed but not btree_gin
-//	// Results in diff.ExtensionsAdded = ["btree_gin"]
-//
-//	// Using custom ignore options
-//	opts := config.WithIgnoredExtensions("plpgsql", "adminpack")
-//	Extensions(generated, database, diff, opts)
-//
-// # Output Consistency
-//
-// Results are sorted alphabetically for consistent output across multiple runs,
-// ensuring deterministic migration generation and reliable testing.
-func Extensions(generated *goschema.Database, database *types.DBSchema, diff *difftypes.SchemaDiff, opts *config.CompareOptions) {
-	// Use default options if none provided
-	if opts == nil {
-		opts = config.DefaultCompareOptions()
-	}
-
-	// Initialize slices to ensure they're never nil
-	diff.ExtensionsAdded = []string{}
-	diff.ExtensionsRemoved = []string{}
-
-	// Create maps for quick lookup, filtering out ignored extensions
-	genExtensions := make(map[string]goschema.Extension)
-	for _, extension := range generated.Extensions {
-		if !opts.IsExtensionIgnored(extension.Name) {
-			genExtensions[extension.Name] = extension
-		}
-	}
-
-	// Create map of database extensions for efficient lookup, filtering out ignored extensions
-	dbExtensions := make(map[string]types.DBExtension)
-	for _, extension := range database.Extensions {
-		if !opts.IsExtensionIgnored(extension.Name) {
-			dbExtensions[extension.Name] = extension
-		}
-	}
-
-	// Find added extensions (exist in generated schema but not in database)
-	// Note: Ignored extensions are already filtered out, so they won't appear here
-	for extensionName := range genExtensions {
-		if _, exists := dbExtensions[extensionName]; !exists {
-			diff.ExtensionsAdded = append(diff.ExtensionsAdded, extensionName)
-		}
-	}
-
-	// Find removed extensions (exist in database but not in generated schema)
-	// Note: Ignored extensions are already filtered out, so they will never be marked for removal
-	for extensionName := range dbExtensions {
-		if _, exists := genExtensions[extensionName]; !exists {
-			diff.ExtensionsRemoved = append(diff.ExtensionsRemoved, extensionName)
-		}
-	}
-
-	// Sort for consistent output
-	sort.Strings(diff.ExtensionsAdded)
-	sort.Strings(diff.ExtensionsRemoved)
-}
-
-// FunctionDefinitions performs detailed comparison between generated and database function definitions.
-//
-// This function compares all aspects of a PostgreSQL function definition to determine
-// if the function needs to be recreated due to changes in its definition. PostgreSQL
-// functions typically require dropping and recreating when modified.
-//
-// # Function Properties Compared
-//
-// The function compares the following properties:
-//   - **Parameters**: Function parameter list and types
-//   - **Returns**: Return type specification
-//   - **Language**: Function language (plpgsql, sql, etc.)
-//   - **Security**: Security context (DEFINER vs INVOKER)
-//   - **Volatility**: Function volatility (STABLE, IMMUTABLE, VOLATILE)
-//   - **Body**: Function implementation code
-//
-// # Example Scenarios
-//
-// **Parameter change**:
-//   - Generated: "get_user_count(tenant_id TEXT)"
-//   - Database: "get_user_count()"
-//   - Result: Changes["parameters"] = "() -> (tenant_id TEXT)"
-//
-// **Body modification**:
-//   - Generated: "SELECT COUNT(*) FROM users WHERE tenant_id = $1"
-//   - Database: "SELECT COUNT(*) FROM users"
-//   - Result: Changes["body"] = "old_body -> new_body"
-//
-// **Volatility change**:
-//   - Generated: STABLE
-//   - Database: VOLATILE
-//   - Result: Changes["volatility"] = "VOLATILE -> STABLE"
-//
-// # Parameters
-//
-//   - genFunction: Generated function definition from Go struct annotations
-//   - dbFunction: Current database function from introspection
-//
-// # Return Value
-//
-// Returns a FunctionDiff containing:
-//   - FunctionName: Name of the function being compared
-//   - Changes: Map of property changes in "old -> new" format
-//
-// # Migration Implications
-//
-// Function changes typically require:
-//  1. DROP FUNCTION (with CASCADE if dependencies exist)
-//  2. CREATE OR REPLACE FUNCTION with new definition
-func FunctionDefinitions(genFunction goschema.Function, dbFunction types.DBFunction) difftypes.FunctionDiff {
-	functionDiff := difftypes.FunctionDiff{
-		FunctionName: genFunction.Name,
-		Changes:      make(map[string]string),
-	}
-
-	// Defense-in-depth: canonicalize a local copy. The annotation parser at
-	// core/goschema/parser.go already calls Canonicalize, but test code (this
-	// package, integration tests) constructs goschema.Function directly with
-	// non-canonical case, and a future programmatic API consumer might too.
-	// The DB-side read path already returns canonical case by construction,
-	// so we only normalize the gen side.
-	genFunction.Canonicalize()
-
-	// Compare parameters
-	if genFunction.Parameters != dbFunction.Parameters {
-		functionDiff.Changes["parameters"] = fmt.Sprintf("%s -> %s", dbFunction.Parameters, genFunction.Parameters)
-	}
-
-	// Compare return type
-	if genFunction.Returns != dbFunction.Returns {
-		functionDiff.Changes["returns"] = fmt.Sprintf("%s -> %s", dbFunction.Returns, genFunction.Returns)
-	}
-
-	// Compare language
-	if genFunction.Language != dbFunction.Language {
-		functionDiff.Changes["language"] = fmt.Sprintf("%s -> %s", dbFunction.Language, genFunction.Language)
-	}
-
-	// Compare security context (DEFINER vs INVOKER)
-	if genFunction.Security != dbFunction.Security {
-		functionDiff.Changes["security"] = fmt.Sprintf("%s -> %s", dbFunction.Security, genFunction.Security)
-	}
-
-	// Compare volatility (VOLATILE/STABLE/IMMUTABLE)
-	if genFunction.Volatility != dbFunction.Volatility {
-		functionDiff.Changes["volatility"] = fmt.Sprintf("%s -> %s", dbFunction.Volatility, genFunction.Volatility)
-	}
-
-	// Compare function body (normalize whitespace for comparison)
-	genBody := strings.TrimSpace(genFunction.Body)
-	dbBody := strings.TrimSpace(dbFunction.Body)
-	if genBody != dbBody {
-		functionDiff.Changes["body"] = fmt.Sprintf("%s -> %s", dbBody, genBody)
-	}
-
-	return functionDiff
-}
-
-// ViewDefinitions performs detailed comparison between generated and database view definitions.
-func ViewDefinitions(genView goschema.View, dbView types.DBView) difftypes.ViewDiff {
-	viewDiff := difftypes.ViewDiff{
-		ViewName: genView.Name,
-		Changes:  make(map[string]string),
-	}
-
-	if !schemaObjectBodiesEqual(genView.Body, dbView.Body) {
-		viewDiff.Changes["body"] = fmt.Sprintf("%s -> %s", strings.TrimSpace(dbView.Body), strings.TrimSpace(genView.Body))
-	}
-
-	dbWithCheck := !strings.EqualFold(dbView.CheckOption, "") && !strings.EqualFold(dbView.CheckOption, "NONE")
-	if genView.WithCheck != dbWithCheck {
-		viewDiff.Changes["with_check"] = fmt.Sprintf("%t -> %t", dbWithCheck, genView.WithCheck)
-	}
-
-	return viewDiff
-}
-
-// MaterializedViewDefinitions performs detailed comparison between generated
-// and database materialized view definitions.
-func MaterializedViewDefinitions(genView goschema.MaterializedView, dbView types.DBMatView) difftypes.MaterializedViewDiff {
-	viewDiff := difftypes.MaterializedViewDiff{
-		ViewName: genView.Name,
-		Changes:  make(map[string]string),
-	}
-
-	if !schemaObjectBodiesEqual(genView.Body, dbView.Body) {
-		viewDiff.Changes["body"] = fmt.Sprintf("%s -> %s", strings.TrimSpace(dbView.Body), strings.TrimSpace(genView.Body))
-	}
-
-	return viewDiff
-}
-
-// TriggerDefinitions performs detailed comparison between generated and database trigger definitions.
-func TriggerDefinitions(genTrigger goschema.Trigger, dbTrigger types.DBTrigger) difftypes.TriggerDiff {
-	genTrigger.Canonicalize()
-
-	triggerDiff := difftypes.TriggerDiff{
-		TriggerName: genTrigger.Name,
-		TableName:   genTrigger.Table,
-		Changes:     make(map[string]string),
-	}
-
-	if genTrigger.Timing != strings.ToUpper(dbTrigger.Timing) {
-		triggerDiff.Changes["timing"] = fmt.Sprintf("%s -> %s", dbTrigger.Timing, genTrigger.Timing)
-	}
-	if genTrigger.Event != strings.ToUpper(dbTrigger.Event) {
-		triggerDiff.Changes["event"] = fmt.Sprintf("%s -> %s", dbTrigger.Event, genTrigger.Event)
-	}
-	dbForEach := strings.ToUpper(strings.TrimSpace(dbTrigger.ForEach))
-	if dbForEach == "" {
-		dbForEach = "ROW"
-	}
-	if genTrigger.ForEach != dbForEach {
-		triggerDiff.Changes["for"] = fmt.Sprintf("%s -> %s", dbForEach, genTrigger.ForEach)
-	}
-
-	genBody := normalizeTriggerBody(genTrigger.Body)
-	dbBody := normalizeTriggerBody(dbTrigger.Body)
-	if genBody != dbBody {
-		triggerDiff.Changes["body"] = fmt.Sprintf("%s -> %s", strings.TrimSpace(dbTrigger.Body), strings.TrimSpace(genTrigger.Body))
-	}
-
-	return triggerDiff
-}
-
-func schemaObjectBodiesEqual(generatedBody, databaseBody string) bool {
-	if normalizeSQLBodyPreservingQualifiers(generatedBody) == normalizeSQLBodyPreservingQualifiers(databaseBody) {
-		return true
-	}
-
-	if schemaQualifierPattern.MatchString(strings.ToLower(generatedBody)) {
-		return false
-	}
-	return normalizeSQLBodyPreservingQualifiers(generatedBody) == normalizeSQLBodyStrippingQualifiers(databaseBody)
-}
-
-func normalizeTriggerBody(body string) string {
-	body = normalizeSQLBodyPreservingQualifiers(body)
-	body = strings.TrimPrefix(body, "begin ")
-	body = strings.TrimPrefix(body, "begin")
-	body = strings.TrimSpace(body)
-	body = strings.TrimSuffix(body, " end")
-	body = strings.TrimSuffix(body, "end")
-	body = strings.TrimSpace(body)
-	body = strings.TrimSuffix(body, ";")
-	return strings.TrimSpace(body)
-}
-
-func normalizeSQLBodyPreservingQualifiers(body string) string {
-	return normalizeSQLBody(body)
-}
-
-func normalizeSQLBodyStrippingQualifiers(body string) string {
-	return schemaQualifierPattern.ReplaceAllString(normalizeSQLBody(body), "")
-}
-
-func normalizeSQLBody(body string) string {
-	body = strings.TrimSpace(body)
-	body = strings.TrimSuffix(body, ";")
-	body = strings.TrimSpace(body)
-	body = strings.ReplaceAll(body, "\"", "")
-	body = strings.ReplaceAll(body, "`", "")
-	body = strings.ToLower(body)
-	body = stripDefaultAggregateAliases(body)
-	body = regexp.MustCompile(`\s+`).ReplaceAllString(body, " ")
-	return strings.TrimSpace(body)
-}
-
-func stripDefaultAggregateAliases(body string) string {
-	return defaultAggregateAliasPattern.ReplaceAllStringFunc(body, func(match string) string {
-		parts := defaultAggregateAliasPattern.FindStringSubmatch(match)
-		if len(parts) != 4 || parts[1] != parts[3] {
-			return match
-		}
-		return parts[1] + "(" + parts[2] + ")"
-	})
-}
-
-// RLSPolicyDefinitions performs detailed comparison between generated and database RLS policy definitions.
-//
-// This function compares all aspects of a PostgreSQL RLS policy definition to determine
-// if the policy needs to be recreated due to changes in its definition. PostgreSQL
-// RLS policies typically require dropping and recreating when modified.
-//
-// # Policy Properties Compared
-//
-// The function compares the following properties:
-//   - **PolicyFor**: Policy type (ALL, SELECT, INSERT, UPDATE, DELETE)
-//   - **ToRoles**: Target database roles
-//   - **UsingExpression**: USING clause for row filtering
-//   - **WithCheckExpression**: WITH CHECK clause for INSERT/UPDATE validation
-//
-// # Example Scenarios
-//
-// **USING expression change**:
-//   - Generated: "tenant_id = get_current_tenant_id()"
-//   - Database: "tenant_id = current_user_id()"
-//   - Result: Changes["using_expression"] = "old_expr -> new_expr"
-//
-// **Role change**:
-//   - Generated: "app_user,admin_user"
-//   - Database: "app_user"
-//   - Result: Changes["to_roles"] = "app_user -> app_user,admin_user"
-//
-// **Policy type change**:
-//   - Generated: "ALL"
-//   - Database: "SELECT"
-//   - Result: Changes["policy_for"] = "SELECT -> ALL"
-//
-// # Parameters
-//
-//   - genPolicy: Generated RLS policy definition from Go struct annotations
-//   - dbPolicy: Current database RLS policy from introspection
-//
-// # Return Value
-//
-// Returns an RLSPolicyDiff containing:
-//   - PolicyName: Name of the policy being compared
-//   - TableName: Name of the table the policy applies to
-//   - Changes: Map of property changes in "old -> new" format
-//
-// # Migration Implications
-//
-// Policy changes typically require:
-//  1. DROP POLICY policy_name ON table_name
-//  2. CREATE POLICY policy_name ON table_name with new definition
-func RLSPolicyDefinitions(genPolicy goschema.RLSPolicy, dbPolicy types.DBRLSPolicy) difftypes.RLSPolicyDiff {
-	policyDiff := difftypes.RLSPolicyDiff{
-		PolicyName: genPolicy.Name,
-		TableName:  genPolicy.Table,
-		Changes:    make(map[string]string),
-	}
-
-	// Compare policy type (FOR clause)
-	if genPolicy.PolicyFor != dbPolicy.PolicyFor {
-		policyDiff.Changes["policy_for"] = fmt.Sprintf("%s -> %s", dbPolicy.PolicyFor, genPolicy.PolicyFor)
-	}
-
-	// Compare target roles (TO clause)
-	if genPolicy.ToRoles != dbPolicy.ToRoles {
-		policyDiff.Changes["to_roles"] = fmt.Sprintf("%s -> %s", dbPolicy.ToRoles, genPolicy.ToRoles)
-	}
-
-	// Compare USING expression
-	if normalize.Expression(genPolicy.UsingExpression) != normalize.Expression(dbPolicy.UsingExpression) {
-		policyDiff.Changes["using_expression"] = fmt.Sprintf("%s -> %s", dbPolicy.UsingExpression, genPolicy.UsingExpression)
-	}
-
-	// Compare WITH CHECK expression
-	if normalize.Expression(genPolicy.WithCheckExpression) != normalize.Expression(dbPolicy.WithCheckExpression) {
-		policyDiff.Changes["with_check_expression"] = fmt.Sprintf("%s -> %s", dbPolicy.WithCheckExpression, genPolicy.WithCheckExpression)
-	}
-
-	return policyDiff
-}
